initial rationalization of engine.go
- breakout params to params.go
  - k8s-specific in params_k8s.go
- breakout template generator to template_generator.go
- move constants to const.go
- move type definitions to types.go
- make unnecessary exportable properties non-exportable
