+++ pkg/acsengine/engine.go
@@ -5,7 +5,6 @@ import (
 	"compress/gzip"
 	"encoding/base64"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"hash/fnv"
 	"io/ioutil"
@@ -14,8 +13,6 @@ import (
 	"net"
 	"net/http"
 	"regexp"
-	"runtime/debug"
-	"sort"
 	"strconv"
 	"strings"
 	"text/template"
@@ -24,102 +21,7 @@ import (
 	"github.com/Azure/acs-engine/pkg/api"
 	"github.com/Azure/acs-engine/pkg/api/common"
 	"github.com/Azure/acs-engine/pkg/helpers"
-	"github.com/Azure/acs-engine/pkg/i18n"
-	"github.com/Masterminds/semver"
 	"github.com/ghodss/yaml"
-)
-
-const (
-	kubernetesMasterCustomDataYaml           = "k8s/kubernetesmastercustomdata.yml"
-	kubernetesCustomScript                   = "k8s/kubernetescustomscript.sh"
-	kubernetesProvisionSourceScript          = "k8s/kubernetesprovisionsource.sh"
-	kubernetesMountetcd                      = "k8s/kubernetes_mountetcd.sh"
-	kubernetesCustomSearchDomainsScript      = "k8s/setup-custom-search-domains.sh"
-	kubernetesMasterGenerateProxyCertsScript = "k8s/kubernetesmastergenerateproxycertscript.sh"
-	kubernetesAgentCustomDataYaml            = "k8s/kubernetesagentcustomdata.yml"
-	kubernetesJumpboxCustomDataYaml          = "k8s/kubernetesjumpboxcustomdata.yml"
-	kubeConfigJSON                           = "k8s/kubeconfig.json"
-	kubernetesWindowsAgentCustomDataPS1      = "k8s/kuberneteswindowssetup.ps1"
-	// OpenShift custom scripts
-	openshiftNodeScript     = "openshift/unstable/openshiftnodescript.sh"
-	openshiftMasterScript   = "openshift/unstable/openshiftmasterscript.sh"
-	openshift39NodeScript   = "openshift/release-3.9/openshiftnodescript.sh"
-	openshift39MasterScript = "openshift/release-3.9/openshiftmasterscript.sh"
-)
-
-const (
-	dcosCustomData188    = "dcos/dcoscustomdata188.t"
-	dcosCustomData190    = "dcos/dcoscustomdata190.t"
-	dcosCustomData198    = "dcos/dcoscustomdata198.t"
-	dcosCustomData110    = "dcos/dcoscustomdata110.t"
-	dcosProvision        = "dcos/dcosprovision.sh"
-	dcosWindowsProvision = "dcos/dcosWindowsProvision.ps1"
-	dcosProvisionSource  = "dcos/dcosprovisionsource.sh"
-
-	dcos2Provision          = "dcos/bstrap/dcosprovision.sh"
-	dcos2BootstrapProvision = "dcos/bstrap/bootstrapprovision.sh"
-	dcos2CustomData1110     = "dcos/bstrap/dcos1.11.0.customdata.t"
-	dcos2CustomData1112     = "dcos/bstrap/dcos1.11.2.customdata.t"
-)
-
-const (
-	swarmProvision        = "swarm/configure-swarm-cluster.sh"
-	swarmWindowsProvision = "swarm/Install-ContainerHost-And-Join-Swarm.ps1"
-
-	swarmModeProvision        = "swarm/configure-swarmmode-cluster.sh"
-	swarmModeWindowsProvision = "swarm/Join-SwarmMode-cluster.ps1"
-)
-
-const (
-	agentOutputs                  = "agentoutputs.t"
-	agentParams                   = "agentparams.t"
-	classicParams                 = "classicparams.t"
-	dcosAgentResourcesVMAS        = "dcos/dcosagentresourcesvmas.t"
-	dcosWindowsAgentResourcesVMAS = "dcos/dcosWindowsAgentResourcesVmas.t"
-	dcosAgentResourcesVMSS        = "dcos/dcosagentresourcesvmss.t"
-	dcosWindowsAgentResourcesVMSS = "dcos/dcosWindowsAgentResourcesVmss.t"
-	dcosAgentVars                 = "dcos/dcosagentvars.t"
-	dcosBaseFile                  = "dcos/dcosbase.t"
-	dcosParams                    = "dcos/dcosparams.t"
-	dcosMasterResources           = "dcos/dcosmasterresources.t"
-	dcosMasterVars                = "dcos/dcosmastervars.t"
-	dcos2BaseFile                 = "dcos/bstrap/dcosbase.t"
-	dcos2BootstrapVars            = "dcos/bstrap/bootstrapvars.t"
-	dcos2BootstrapParams          = "dcos/bstrap/bootstrapparams.t"
-	dcos2BootstrapResources       = "dcos/bstrap/bootstrapresources.t"
-	dcos2BootstrapCustomdata      = "dcos/bstrap/bootstrapcustomdata.yml"
-	dcos2MasterVars               = "dcos/bstrap/dcosmastervars.t"
-	dcos2MasterResources          = "dcos/bstrap/dcosmasterresources.t"
-	iaasOutputs                   = "iaasoutputs.t"
-	kubernetesBaseFile            = "k8s/kubernetesbase.t"
-	kubernetesAgentResourcesVMAS  = "k8s/kubernetesagentresourcesvmas.t"
-	kubernetesAgentResourcesVMSS  = "k8s/kubernetesagentresourcesvmss.t"
-	kubernetesAgentVars           = "k8s/kubernetesagentvars.t"
-	kubernetesMasterResources     = "k8s/kubernetesmasterresources.t"
-	kubernetesMasterVars          = "k8s/kubernetesmastervars.t"
-	kubernetesParams              = "k8s/kubernetesparams.t"
-	kubernetesWinAgentVars        = "k8s/kuberneteswinagentresourcesvmas.t"
-	kubernetesWinAgentVarsVMSS    = "k8s/kuberneteswinagentresourcesvmss.t"
-	masterOutputs                 = "masteroutputs.t"
-	masterParams                  = "masterparams.t"
-	swarmBaseFile                 = "swarm/swarmbase.t"
-	swarmParams                   = "swarm/swarmparams.t"
-	swarmAgentResourcesVMAS       = "swarm/swarmagentresourcesvmas.t"
-	swarmAgentResourcesVMSS       = "swarm/swarmagentresourcesvmss.t"
-	swarmAgentResourcesClassic    = "swarm/swarmagentresourcesclassic.t"
-	swarmAgentVars                = "swarm/swarmagentvars.t"
-	swarmMasterResources          = "swarm/swarmmasterresources.t"
-	swarmMasterVars               = "swarm/swarmmastervars.t"
-	swarmWinAgentResourcesVMAS    = "swarm/swarmwinagentresourcesvmas.t"
-	swarmWinAgentResourcesVMSS    = "swarm/swarmwinagentresourcesvmss.t"
-	windowsParams                 = "windowsparams.t"
-)
-
-const (
-	azurePublicCloud       = "AzurePublicCloud"
-	azureChinaCloud        = "AzureChinaCloud"
-	azureGermanCloud       = "AzureGermanCloud"
-	azureUSGovernmentCloud = "AzureUSGovernmentCloud"
 )
 
 var commonTemplateFiles = []string{agentOutputs, agentParams, classicParams, masterOutputs, iaasOutputs, masterParams, windowsParams}
@@ -136,193 +38,10 @@ var openshiftTemplateFiles = append(
 	openshift39MasterScript,
 )
 
-func getOpenshiftMasterShAsset(version string) string {
-	switch version {
-	case common.OpenShiftVersion3Dot9Dot0:
-		return openshift39MasterScript
-	case common.OpenShiftVersionUnstable:
-		return openshiftMasterScript
-	default:
-		panic(fmt.Sprintf("BUG: invalid OpenShift version %s", version))
-	}
-}
-
-func getOpenshiftNodeShAsset(version string) string {
-	switch version {
-	case common.OpenShiftVersion3Dot9Dot0:
-		return openshift39NodeScript
-	case common.OpenShiftVersionUnstable:
-		return openshiftNodeScript
-	default:
-		panic(fmt.Sprintf("BUG: invalid OpenShift version %s", version))
-	}
-}
-
-/**
- The following parameters could be either a plain text, or referenced to a secret in a keyvault:
- - apiServerCertificate
- - apiServerPrivateKey
- - caCertificate
- - clientCertificate
- - clientPrivateKey
- - kubeConfigCertificate
- - kubeConfigPrivateKey
- - servicePrincipalClientSecret
- - etcdClientCertificate
- - etcdClientPrivateKey
- - etcdServerCertificate
- - etcdServerPrivateKey
- - etcdPeerCertificates
- - etcdPeerPrivateKeys
-
- To refer to a keyvault secret, the value of the parameter in the api model file should be formatted as:
-
- "<PARAMETER>": "/subscriptions/<SUB_ID>/resourceGroups/<RG_NAME>/providers/Microsoft.KeyVault/vaults/<KV_NAME>/secrets/<NAME>[/<VERSION>]"
- where:
-   <SUB_ID> is the subscription ID of the keyvault
-   <RG_NAME> is the resource group of the keyvault
-   <KV_NAME> is the name of the keyvault
-   <NAME> is the name of the secret.
-   <VERSION> (optional) is the version of the secret (default: the latest version)
-
- This will generate a reference block in the parameters file:
-
- "reference": {
-   "keyVault": {
-     "id": "/subscriptions/<SUB_ID>/resourceGroups/<RG_NAME>/providers/Microsoft.KeyVault/vaults/<KV_NAME>"
-   },
-   "secretName": "<NAME>"
-   "secretVersion": "<VERSION>"
-}
-**/
-
-// KeyVaultID represents a KeyVault instance on Azure
-type KeyVaultID struct {
-	ID string `json:"id"`
-}
-
-// KeyVaultRef represents a reference to KeyVault instance on Azure
-type KeyVaultRef struct {
-	KeyVault      KeyVaultID `json:"keyVault"`
-	SecretName    string     `json:"secretName"`
-	SecretVersion string     `json:"secretVersion,omitempty"`
-}
-
-type paramsMap map[string]interface{}
-
 var keyvaultSecretPathRe *regexp.Regexp
 
 func init() {
 	keyvaultSecretPathRe = regexp.MustCompile(`^(/subscriptions/\S+/resourceGroups/\S+/providers/Microsoft.KeyVault/vaults/\S+)/secrets/([^/\s]+)(/(\S+))?$`)
-}
-
-func (t *TemplateGenerator) verifyFiles() error {
-	allFiles := commonTemplateFiles
-	allFiles = append(allFiles, dcosTemplateFiles...)
-	allFiles = append(allFiles, dcos2TemplateFiles...)
-	allFiles = append(allFiles, kubernetesTemplateFiles...)
-	allFiles = append(allFiles, openshiftTemplateFiles...)
-	allFiles = append(allFiles, swarmTemplateFiles...)
-	for _, file := range allFiles {
-		if _, err := Asset(file); err != nil {
-			return t.Translator.Errorf("template file %s does not exist", file)
-		}
-	}
-	return nil
-}
-
-// TemplateGenerator represents the object that performs the template generation.
-type TemplateGenerator struct {
-	ClassicMode bool
-	Translator  *i18n.Translator
-}
-
-// InitializeTemplateGenerator creates a new template generator object
-func InitializeTemplateGenerator(ctx Context, classicMode bool) (*TemplateGenerator, error) {
-	t := &TemplateGenerator{
-		ClassicMode: classicMode,
-		Translator:  ctx.Translator,
-	}
-
-	if err := t.verifyFiles(); err != nil {
-		return nil, err
-	}
-
-	return t, nil
-}
-
-// GenerateTemplate generates the template from the API Model
-func (t *TemplateGenerator) GenerateTemplate(containerService *api.ContainerService, generatorCode string, isUpgrade bool, acsengineVersion string) (templateRaw string, parametersRaw string, certsGenerated bool, err error) {
-	// named return values are used in order to set err in case of a panic
-	templateRaw = ""
-	parametersRaw = ""
-	err = nil
-
-	var templ *template.Template
-
-	properties := containerService.Properties
-	// save the current orchestrator version and restore it after deploying.
-	// this allows us to deploy agents on the most recent patch without updating the orchestator version in the object
-	orchVersion := properties.OrchestratorProfile.OrchestratorVersion
-	defer func() {
-		properties.OrchestratorProfile.OrchestratorVersion = orchVersion
-	}()
-	if certsGenerated, err = SetPropertiesDefaults(containerService, isUpgrade); err != nil {
-		return templateRaw, parametersRaw, certsGenerated, err
-	}
-
-	templ = template.New("acs template").Funcs(t.getTemplateFuncMap(containerService))
-
-	files, baseFile, e := t.prepareTemplateFiles(properties)
-	if e != nil {
-		return "", "", false, e
-	}
-
-	for _, file := range files {
-		bytes, e := Asset(file)
-		if e != nil {
-			err = t.Translator.Errorf("Error reading file %s, Error: %s", file, e.Error())
-			return templateRaw, parametersRaw, certsGenerated, err
-		}
-		if _, err = templ.New(file).Parse(string(bytes)); err != nil {
-			return templateRaw, parametersRaw, certsGenerated, err
-		}
-	}
-	// template generation may have panics in the called functions.  This catches those panics
-	// and ensures the panic is returned as an error
-	defer func() {
-		if r := recover(); r != nil {
-			s := debug.Stack()
-			err = fmt.Errorf("%v - %s", r, s)
-
-			// invalidate the template and the parameters
-			templateRaw = ""
-			parametersRaw = ""
-		}
-	}()
-
-	if !ValidateDistro(containerService) {
-		return templateRaw, parametersRaw, certsGenerated, fmt.Errorf("Invalid distro")
-	}
-
-	var b bytes.Buffer
-	if err = templ.ExecuteTemplate(&b, baseFile, properties); err != nil {
-		return templateRaw, parametersRaw, certsGenerated, err
-	}
-	templateRaw = b.String()
-
-	var parametersMap paramsMap
-	if parametersMap, err = getParameters(containerService, t.ClassicMode, generatorCode, acsengineVersion); err != nil {
-		return templateRaw, parametersRaw, certsGenerated, err
-	}
-
-	var parameterBytes []byte
-	if parameterBytes, err = helpers.JSONMarshal(parametersMap, false); err != nil {
-		return templateRaw, parametersRaw, certsGenerated, err
-	}
-	parametersRaw = string(parameterBytes)
-
-	return templateRaw, parametersRaw, certsGenerated, err
 }
 
 // GenerateClusterID creates a unique 8 string cluster ID
@@ -381,7 +100,7 @@ func GenerateKubeConfig(properties *api.Properties, location string) (string, er
 		}
 
 		authInfo = fmt.Sprintf("{\"auth-provider\":{\"name\":\"azure\",\"config\":{\"environment\":\"%v\",\"tenant-id\":\"%v\",\"apiserver-id\":\"%v\",\"client-id\":\"%v\"}}}",
-			GetCloudTargetEnv(location),
+			getCloudTargetEnv(location),
 			tenantID,
 			properties.AADProfile.ServerAppID,
 			properties.AADProfile.ClientAppID)
@@ -391,51 +110,8 @@ func GenerateKubeConfig(properties *api.Properties, location string) (string, er
 	return kubeconfig, nil
 }
 
-func generateIPList(count int, firstAddr string) []string {
-	ipaddr := net.ParseIP(firstAddr).To4()
-	if ipaddr == nil {
-		panic(fmt.Sprintf("IPAddr '%s' is an invalid IP address", firstAddr))
-	}
-	ret := make([]string, count)
-	for i := 0; i < count; i++ {
-		ret[i] = fmt.Sprintf("%d.%d.%d.%d", ipaddr[0], ipaddr[1], ipaddr[2], ipaddr[3]+byte(i))
-	}
-	return ret
-}
-
-func (t *TemplateGenerator) prepareTemplateFiles(properties *api.Properties) ([]string, string, error) {
-	var files []string
-	var baseFile string
-	switch properties.OrchestratorProfile.OrchestratorType {
-	case api.DCOS:
-		if properties.OrchestratorProfile.DcosConfig == nil || properties.OrchestratorProfile.DcosConfig.BootstrapProfile == nil {
-			files = append(commonTemplateFiles, dcosTemplateFiles...)
-			baseFile = dcosBaseFile
-		} else {
-			files = append(commonTemplateFiles, dcos2TemplateFiles...)
-			baseFile = dcos2BaseFile
-		}
-	case api.Swarm:
-		files = append(commonTemplateFiles, swarmTemplateFiles...)
-		baseFile = swarmBaseFile
-	case api.Kubernetes:
-		files = append(commonTemplateFiles, kubernetesTemplateFiles...)
-		baseFile = kubernetesBaseFile
-	case api.SwarmMode:
-		files = append(commonTemplateFiles, swarmModeTemplateFiles...)
-		baseFile = swarmBaseFile
-	case api.OpenShift:
-		files = append(commonTemplateFiles, openshiftTemplateFiles...)
-		baseFile = kubernetesBaseFile
-	default:
-		return nil, "", t.Translator.Errorf("orchestrator '%s' is unsupported", properties.OrchestratorProfile.OrchestratorType)
-	}
-
-	return files, baseFile, nil
-}
-
-// FormatAzureProdFQDNs constructs all possible Azure prod fqdn
-func FormatAzureProdFQDNs(fqdnPrefix string) []string {
+// formatAzureProdFQDNs constructs all possible Azure prod fqdn
+func formatAzureProdFQDNs(fqdnPrefix string) []string {
 	var fqdns []string
 	for _, location := range AzureLocations {
 		fqdns = append(fqdns, FormatAzureProdFQDN(fqdnPrefix, location))
@@ -446,7 +122,7 @@ func FormatAzureProdFQDNs(fqdnPrefix string) []string {
 // FormatAzureProdFQDN constructs an Azure prod fqdn
 func FormatAzureProdFQDN(fqdnPrefix string, location string) string {
 	var FQDNFormat string
-	switch GetCloudTargetEnv(location) {
+	switch getCloudTargetEnv(location) {
 	case azureChinaCloud:
 		FQDNFormat = AzureChinaCloudSpec.EndpointConfig.ResourceManagerVMDNSSuffix
 	case azureGermanCloud:
@@ -459,11 +135,11 @@ func FormatAzureProdFQDN(fqdnPrefix string, location string) string {
 	return fmt.Sprintf("%s.%s."+FQDNFormat, fqdnPrefix, location)
 }
 
-//GetCloudSpecConfig returns the kubenernetes container images url configurations based on the deploy target environment
+//getCloudSpecConfig returns the kubenernetes container images url configurations based on the deploy target environment
 //for example: if the target is the public azure, then the default container image url should be k8s-gcrio.azureedge.net/...
 //if the target is azure china, then the default container image should be mirror.azure.cn:5000/google_container/...
-func GetCloudSpecConfig(location string) AzureEnvironmentSpecConfig {
-	switch GetCloudTargetEnv(location) {
+func getCloudSpecConfig(location string) AzureEnvironmentSpecConfig {
+	switch getCloudTargetEnv(location) {
 	case azureChinaCloud:
 		return AzureChinaCloudSpec
 	case azureGermanCloud:
@@ -475,8 +151,8 @@ func GetCloudSpecConfig(location string) AzureEnvironmentSpecConfig {
 	}
 }
 
-// ValidateDistro checks if the requested orchestrator type is supported on the requested Linux distro.
-func ValidateDistro(cs *api.ContainerService) bool {
+// validateDistro checks if the requested orchestrator type is supported on the requested Linux distro.
+func validateDistro(cs *api.ContainerService) bool {
 	// Check Master distro
 	if cs.Properties.MasterProfile != nil && cs.Properties.MasterProfile.Distro == api.RHEL &&
 		(cs.Properties.OrchestratorProfile.OrchestratorType != api.SwarmMode && cs.Properties.OrchestratorProfile.OrchestratorType != api.OpenShift) {
@@ -494,10 +170,10 @@ func ValidateDistro(cs *api.ContainerService) bool {
 	return true
 }
 
-// GetCloudTargetEnv determines and returns whether the region is a sovereign cloud which
+// getCloudTargetEnv determines and returns whether the region is a sovereign cloud which
 // have their own data compliance regulations (China/Germany/USGov) or standard
 //  Azure public cloud
-func GetCloudTargetEnv(location string) string {
+func getCloudTargetEnv(location string) string {
 	loc := strings.ToLower(strings.Join(strings.Fields(location), ""))
 	switch {
 	case loc == "chinaeast" || loc == "chinanorth":
@@ -511,463 +187,38 @@ func GetCloudTargetEnv(location string) string {
 	}
 }
 
-func getParameters(cs *api.ContainerService, isClassicMode bool, generatorCode string, acsengineVersion string) (paramsMap, error) {
-	properties := cs.Properties
-	location := cs.Location
-	parametersMap := paramsMap{}
-	cloudSpecConfig := GetCloudSpecConfig(location)
-
-	// acsengine Parameters
-	addValue(parametersMap, "acsengineVersion", acsengineVersion)
-
-	// Master Parameters
-	addValue(parametersMap, "location", location)
-
-	// Identify Master distro
-	masterDistro := getMasterDistro(properties.MasterProfile)
-	if properties.MasterProfile != nil && properties.MasterProfile.ImageRef != nil {
-		addValue(parametersMap, "osImageName", properties.MasterProfile.ImageRef.Name)
-		addValue(parametersMap, "osImageResourceGroup", properties.MasterProfile.ImageRef.ResourceGroup)
+func getOpenshiftMasterShAsset(version string) string {
+	switch version {
+	case common.OpenShiftVersion3Dot9Dot0:
+		return openshift39MasterScript
+	case common.OpenShiftVersionUnstable:
+		return openshiftMasterScript
+	default:
+		panic(fmt.Sprintf("BUG: invalid OpenShift version %s", version))
 	}
-	// TODO: Choose the correct image config based on the version
-	// for the openshift orchestrator
-	addValue(parametersMap, "osImageOffer", cloudSpecConfig.OSImageConfig[masterDistro].ImageOffer)
-	addValue(parametersMap, "osImageSKU", cloudSpecConfig.OSImageConfig[masterDistro].ImageSku)
-	addValue(parametersMap, "osImagePublisher", cloudSpecConfig.OSImageConfig[masterDistro].ImagePublisher)
-	addValue(parametersMap, "osImageVersion", cloudSpecConfig.OSImageConfig[masterDistro].ImageVersion)
+}
 
-	addValue(parametersMap, "fqdnEndpointSuffix", cloudSpecConfig.EndpointConfig.ResourceManagerVMDNSSuffix)
-	addValue(parametersMap, "targetEnvironment", GetCloudTargetEnv(location))
-	addValue(parametersMap, "linuxAdminUsername", properties.LinuxProfile.AdminUsername)
-	if properties.LinuxProfile.CustomSearchDomain != nil {
-		addValue(parametersMap, "searchDomainName", properties.LinuxProfile.CustomSearchDomain.Name)
-		addValue(parametersMap, "searchDomainRealmUser", properties.LinuxProfile.CustomSearchDomain.RealmUser)
-		addValue(parametersMap, "searchDomainRealmPassword", properties.LinuxProfile.CustomSearchDomain.RealmPassword)
+func getOpenshiftNodeShAsset(version string) string {
+	switch version {
+	case common.OpenShiftVersion3Dot9Dot0:
+		return openshift39NodeScript
+	case common.OpenShiftVersionUnstable:
+		return openshiftNodeScript
+	default:
+		panic(fmt.Sprintf("BUG: invalid OpenShift version %s", version))
 	}
-	// masterEndpointDNSNamePrefix is the basis for storage account creation across dcos, swarm, and k8s
-	if properties.MasterProfile != nil {
-		// MasterProfile exists, uses master DNS prefix
-		addValue(parametersMap, "masterEndpointDNSNamePrefix", properties.MasterProfile.DNSPrefix)
-	} else if properties.HostedMasterProfile != nil {
-		// Agents only, use cluster DNS prefix
-		addValue(parametersMap, "masterEndpointDNSNamePrefix", properties.HostedMasterProfile.DNSPrefix)
+}
+
+func generateIPList(count int, firstAddr string) []string {
+	ipaddr := net.ParseIP(firstAddr).To4()
+	if ipaddr == nil {
+		panic(fmt.Sprintf("IPAddr '%s' is an invalid IP address", firstAddr))
 	}
-	if properties.MasterProfile != nil {
-		if properties.MasterProfile.IsCustomVNET() {
-			addValue(parametersMap, "masterVnetSubnetID", properties.MasterProfile.VnetSubnetID)
-			if properties.OrchestratorProfile.IsKubernetes() || properties.OrchestratorProfile.IsOpenShift() {
-				addValue(parametersMap, "vnetCidr", properties.MasterProfile.VnetCidr)
-			}
-		} else {
-			addValue(parametersMap, "masterSubnet", properties.MasterProfile.Subnet)
-		}
-		addValue(parametersMap, "firstConsecutiveStaticIP", properties.MasterProfile.FirstConsecutiveStaticIP)
-		addValue(parametersMap, "masterVMSize", properties.MasterProfile.VMSize)
-		if isClassicMode {
-			addValue(parametersMap, "masterCount", properties.MasterProfile.Count)
-		}
+	ret := make([]string, count)
+	for i := 0; i < count; i++ {
+		ret[i] = fmt.Sprintf("%d.%d.%d.%d", ipaddr[0], ipaddr[1], ipaddr[2], ipaddr[3]+byte(i))
 	}
-	if properties.HostedMasterProfile != nil {
-		addValue(parametersMap, "masterSubnet", properties.HostedMasterProfile.Subnet)
-	}
-	addValue(parametersMap, "sshRSAPublicKey", properties.LinuxProfile.SSH.PublicKeys[0].KeyData)
-	for i, s := range properties.LinuxProfile.Secrets {
-		addValue(parametersMap, fmt.Sprintf("linuxKeyVaultID%d", i), s.SourceVault.ID)
-		for j, c := range s.VaultCertificates {
-			addValue(parametersMap, fmt.Sprintf("linuxKeyVaultID%dCertificateURL%d", i, j), c.CertificateURL)
-		}
-	}
-
-	//Swarm and SwarmMode Parameters
-	if properties.OrchestratorProfile.OrchestratorType == api.Swarm || properties.OrchestratorProfile.OrchestratorType == api.SwarmMode {
-		var dockerEngineRepo, dockerComposeDownloadURL string
-		if cloudSpecConfig.DockerSpecConfig.DockerEngineRepo == "" {
-			dockerEngineRepo = DefaultDockerEngineRepo
-		} else {
-			dockerEngineRepo = cloudSpecConfig.DockerSpecConfig.DockerEngineRepo
-		}
-		if cloudSpecConfig.DockerSpecConfig.DockerComposeDownloadURL == "" {
-			dockerComposeDownloadURL = DefaultDockerComposeURL
-		} else {
-			dockerComposeDownloadURL = cloudSpecConfig.DockerSpecConfig.DockerComposeDownloadURL
-		}
-		addValue(parametersMap, "dockerEngineDownloadRepo", dockerEngineRepo)
-		addValue(parametersMap, "dockerComposeDownloadURL", dockerComposeDownloadURL)
-	}
-
-	// Kubernetes Parameters
-	if properties.OrchestratorProfile.IsKubernetes() ||
-		properties.OrchestratorProfile.IsOpenShift() {
-		k8sVersion := properties.OrchestratorProfile.OrchestratorVersion
-
-		kubernetesHyperkubeSpec := properties.OrchestratorProfile.KubernetesConfig.KubernetesImageBase + KubeConfigs[k8sVersion]["hyperkube"]
-		if properties.OrchestratorProfile.KubernetesConfig.CustomHyperkubeImage != "" {
-			kubernetesHyperkubeSpec = properties.OrchestratorProfile.KubernetesConfig.CustomHyperkubeImage
-		}
-
-		dockerEngineVersion := KubeConfigs[k8sVersion]["dockerEngineVersion"]
-		if properties.OrchestratorProfile.KubernetesConfig.DockerEngineVersion != "" {
-			dockerEngineVersion = properties.OrchestratorProfile.KubernetesConfig.DockerEngineVersion
-		}
-
-		if properties.CertificateProfile != nil {
-			addSecret(parametersMap, "apiServerCertificate", properties.CertificateProfile.APIServerCertificate, true)
-			addSecret(parametersMap, "apiServerPrivateKey", properties.CertificateProfile.APIServerPrivateKey, true)
-			addSecret(parametersMap, "caCertificate", properties.CertificateProfile.CaCertificate, true)
-			addSecret(parametersMap, "caPrivateKey", properties.CertificateProfile.CaPrivateKey, true)
-			addSecret(parametersMap, "clientCertificate", properties.CertificateProfile.ClientCertificate, true)
-			addSecret(parametersMap, "clientPrivateKey", properties.CertificateProfile.ClientPrivateKey, true)
-			addSecret(parametersMap, "kubeConfigCertificate", properties.CertificateProfile.KubeConfigCertificate, true)
-			addSecret(parametersMap, "kubeConfigPrivateKey", properties.CertificateProfile.KubeConfigPrivateKey, true)
-			if properties.MasterProfile != nil {
-				addSecret(parametersMap, "etcdServerCertificate", properties.CertificateProfile.EtcdServerCertificate, true)
-				addSecret(parametersMap, "etcdServerPrivateKey", properties.CertificateProfile.EtcdServerPrivateKey, true)
-				addSecret(parametersMap, "etcdClientCertificate", properties.CertificateProfile.EtcdClientCertificate, true)
-				addSecret(parametersMap, "etcdClientPrivateKey", properties.CertificateProfile.EtcdClientPrivateKey, true)
-				for i, pc := range properties.CertificateProfile.EtcdPeerCertificates {
-					addSecret(parametersMap, "etcdPeerCertificate"+strconv.Itoa(i), pc, true)
-				}
-				for i, pk := range properties.CertificateProfile.EtcdPeerPrivateKeys {
-					addSecret(parametersMap, "etcdPeerPrivateKey"+strconv.Itoa(i), pk, true)
-				}
-			}
-		}
-
-		if properties.HostedMasterProfile != nil && properties.HostedMasterProfile.FQDN != "" {
-			addValue(parametersMap, "kubernetesEndpoint", properties.HostedMasterProfile.FQDN)
-		}
-
-		if helpers.IsTrueBoolPointer(properties.OrchestratorProfile.KubernetesConfig.UseCloudControllerManager) {
-			kubernetesCcmSpec := properties.OrchestratorProfile.KubernetesConfig.KubernetesImageBase + KubeConfigs[k8sVersion]["ccm"]
-			if properties.OrchestratorProfile.KubernetesConfig.CustomCcmImage != "" {
-				kubernetesCcmSpec = properties.OrchestratorProfile.KubernetesConfig.CustomCcmImage
-			}
-
-			addValue(parametersMap, "kubernetesCcmImageSpec", kubernetesCcmSpec)
-		}
-
-		if !properties.OrchestratorProfile.IsOpenShift() {
-			addValue(parametersMap, "dockerEngineDownloadRepo", cloudSpecConfig.DockerSpecConfig.DockerEngineRepo)
-			addValue(parametersMap, "dockerEngineVersion", dockerEngineVersion)
-		}
-
-		addValue(parametersMap, "kubeDNSServiceIP", properties.OrchestratorProfile.KubernetesConfig.DNSServiceIP)
-		addValue(parametersMap, "kubeServiceCidr", properties.OrchestratorProfile.KubernetesConfig.ServiceCIDR)
-		addValue(parametersMap, "kubernetesHyperkubeSpec", kubernetesHyperkubeSpec)
-		addValue(parametersMap, "kubernetesAddonManagerSpec", cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase+KubeConfigs[k8sVersion]["addonmanager"])
-		addValue(parametersMap, "kubernetesAddonResizerSpec", cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase+KubeConfigs[k8sVersion]["addonresizer"])
-		addValue(parametersMap, "kubernetesDNSMasqSpec", cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase+KubeConfigs[k8sVersion]["dnsmasq"])
-		addValue(parametersMap, "kubernetesExecHealthzSpec", cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase+KubeConfigs[k8sVersion]["exechealthz"])
-		addValue(parametersMap, "kubernetesHeapsterSpec", cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase+KubeConfigs[k8sVersion]["heapster"])
-		tillerAddon := getAddonByName(properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultTillerAddonName)
-		c := getAddonContainersIndexByName(tillerAddon.Containers, DefaultTillerAddonName)
-		if c > -1 {
-			addValue(parametersMap, "kubernetesTillerCPURequests", tillerAddon.Containers[c].CPURequests)
-			addValue(parametersMap, "kubernetesTillerCPULimit", tillerAddon.Containers[c].CPULimits)
-			addValue(parametersMap, "kubernetesTillerMemoryRequests", tillerAddon.Containers[c].MemoryRequests)
-			addValue(parametersMap, "kubernetesTillerMemoryLimit", tillerAddon.Containers[c].MemoryLimits)
-			addValue(parametersMap, "kubernetesTillerMaxHistory", tillerAddon.Config["max-history"])
-			if tillerAddon.Containers[c].Image != "" {
-				addValue(parametersMap, "kubernetesTillerSpec", tillerAddon.Containers[c].Image)
-			} else {
-				addValue(parametersMap, "kubernetesTillerSpec", cloudSpecConfig.KubernetesSpecConfig.TillerImageBase+KubeConfigs[k8sVersion][DefaultTillerAddonName])
-			}
-		}
-		aciConnectorAddon := getAddonByName(properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultACIConnectorAddonName)
-		c = getAddonContainersIndexByName(aciConnectorAddon.Containers, DefaultACIConnectorAddonName)
-		if c > -1 {
-			addValue(parametersMap, "kubernetesACIConnectorEnabled", aciConnectorAddon.Enabled)
-			addValue(parametersMap, "kubernetesACIConnectorNodeName", aciConnectorAddon.Config["nodeName"])
-			addValue(parametersMap, "kubernetesACIConnectorOS", aciConnectorAddon.Config["os"])
-			addValue(parametersMap, "kubernetesACIConnectorTaint", aciConnectorAddon.Config["taint"])
-			addValue(parametersMap, "kubernetesACIConnectorRegion", aciConnectorAddon.Config["region"])
-			addValue(parametersMap, "kubernetesACIConnectorCPURequests", aciConnectorAddon.Containers[c].CPURequests)
-			addValue(parametersMap, "kubernetesACIConnectorCPULimit", aciConnectorAddon.Containers[c].CPULimits)
-			addValue(parametersMap, "kubernetesACIConnectorMemoryRequests", aciConnectorAddon.Containers[c].MemoryRequests)
-			addValue(parametersMap, "kubernetesACIConnectorMemoryLimit", aciConnectorAddon.Containers[c].MemoryLimits)
-			if aciConnectorAddon.Containers[c].Image != "" {
-				addValue(parametersMap, "kubernetesACIConnectorSpec", aciConnectorAddon.Containers[c].Image)
-			} else {
-				addValue(parametersMap, "kubernetesACIConnectorSpec", cloudSpecConfig.KubernetesSpecConfig.ACIConnectorImageBase+KubeConfigs[k8sVersion][DefaultACIConnectorAddonName])
-			}
-		}
-		clusterAutoscalerAddon := getAddonByName(properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultClusterAutoscalerAddonName)
-		c = getAddonContainersIndexByName(clusterAutoscalerAddon.Containers, DefaultClusterAutoscalerAddonName)
-		if c > -1 {
-			addValue(parametersMap, "kubernetesClusterAutoscalerCPURequests", clusterAutoscalerAddon.Containers[c].CPURequests)
-			addValue(parametersMap, "kubernetesClusterAutoscalerCPULimit", clusterAutoscalerAddon.Containers[c].CPULimits)
-			addValue(parametersMap, "kubernetesClusterAutoscalerMemoryRequests", clusterAutoscalerAddon.Containers[c].MemoryRequests)
-			addValue(parametersMap, "kubernetesClusterAutoscalerMemoryLimit", clusterAutoscalerAddon.Containers[c].MemoryLimits)
-			addValue(parametersMap, "kubernetesClusterAutoscalerMinNodes", clusterAutoscalerAddon.Config["minNodes"])
-			addValue(parametersMap, "kubernetesClusterAutoscalerMaxNodes", clusterAutoscalerAddon.Config["maxNodes"])
-			addValue(parametersMap, "kubernetesClusterAutoscalerEnabled", clusterAutoscalerAddon.Enabled)
-			addValue(parametersMap, "kubernetesClusterAutoscalerUseManagedIdentity", strings.ToLower(strconv.FormatBool(properties.OrchestratorProfile.KubernetesConfig.UseManagedIdentity)))
-			if clusterAutoscalerAddon.Containers[c].Image != "" {
-				addValue(parametersMap, "kubernetesClusterAutoscalerSpec", clusterAutoscalerAddon.Containers[c].Image)
-			} else {
-				addValue(parametersMap, "kubernetesClusterAutoscalerSpec", cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase+KubeConfigs[k8sVersion][DefaultClusterAutoscalerAddonName])
-			}
-		}
-		dashboardAddon := getAddonByName(properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultDashboardAddonName)
-		c = getAddonContainersIndexByName(dashboardAddon.Containers, DefaultDashboardAddonName)
-		if c > -1 {
-			addValue(parametersMap, "kubernetesDashboardCPURequests", dashboardAddon.Containers[c].CPURequests)
-			addValue(parametersMap, "kubernetesDashboardCPULimit", dashboardAddon.Containers[c].CPULimits)
-			addValue(parametersMap, "kubernetesDashboardMemoryRequests", dashboardAddon.Containers[c].MemoryRequests)
-			addValue(parametersMap, "kubernetesDashboardMemoryLimit", dashboardAddon.Containers[c].MemoryLimits)
-			if dashboardAddon.Containers[c].Image != "" {
-				addValue(parametersMap, "kubernetesDashboardSpec", dashboardAddon.Containers[c].Image)
-			} else {
-				addValue(parametersMap, "kubernetesDashboardSpec", cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase+KubeConfigs[k8sVersion][DefaultDashboardAddonName])
-			}
-		}
-		reschedulerAddon := getAddonByName(properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultReschedulerAddonName)
-		c = getAddonContainersIndexByName(reschedulerAddon.Containers, DefaultReschedulerAddonName)
-		if c > -1 {
-			addValue(parametersMap, "kubernetesReschedulerCPURequests", reschedulerAddon.Containers[c].CPURequests)
-			addValue(parametersMap, "kubernetesReschedulerCPULimit", reschedulerAddon.Containers[c].CPULimits)
-			addValue(parametersMap, "kubernetesReschedulerMemoryRequests", reschedulerAddon.Containers[c].MemoryRequests)
-			addValue(parametersMap, "kubernetesReschedulerMemoryLimit", reschedulerAddon.Containers[c].MemoryLimits)
-			if reschedulerAddon.Containers[c].Image != "" {
-				addValue(parametersMap, "kubernetesReschedulerSpec", dashboardAddon.Containers[c].Image)
-			} else {
-				addValue(parametersMap, "kubernetesReschedulerSpec", cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase+KubeConfigs[k8sVersion][DefaultReschedulerAddonName])
-			}
-		}
-		metricsServerAddon := getAddonByName(properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultMetricsServerAddonName)
-		c = getAddonContainersIndexByName(metricsServerAddon.Containers, DefaultMetricsServerAddonName)
-		if c > -1 {
-			if metricsServerAddon.Containers[c].Image != "" {
-				addValue(parametersMap, "kubernetesMetricsServerSpec", metricsServerAddon.Containers[c].Image)
-			} else {
-				addValue(parametersMap, "kubernetesMetricsServerSpec", cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase+KubeConfigs[k8sVersion][DefaultMetricsServerAddonName])
-			}
-		}
-		nvidiaDevicePluginAddon := getAddonByName(properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultNVIDIADevicePluginAddonName)
-		c = getAddonContainersIndexByName(nvidiaDevicePluginAddon.Containers, DefaultNVIDIADevicePluginAddonName)
-		if c > -1 {
-			if nvidiaDevicePluginAddon.Containers[c].Image != "" {
-				addValue(parametersMap, "kubernetesNVIDIADevicePluginSpec", nvidiaDevicePluginAddon.Containers[c].Image)
-			} else {
-				addValue(parametersMap, "kubernetesNVIDIADevicePluginSpec", cloudSpecConfig.KubernetesSpecConfig.NVIDIAImageBase+KubeConfigs[k8sVersion][DefaultNVIDIADevicePluginAddonName])
-			}
-		}
-		addValue(parametersMap, "kubernetesKubeDNSSpec", cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase+KubeConfigs[k8sVersion]["dns"])
-		addValue(parametersMap, "kubernetesPodInfraContainerSpec", cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase+KubeConfigs[k8sVersion]["pause"])
-		addValue(parametersMap, "cloudProviderBackoff", strconv.FormatBool(properties.OrchestratorProfile.KubernetesConfig.CloudProviderBackoff))
-		addValue(parametersMap, "cloudProviderBackoffRetries", strconv.Itoa(properties.OrchestratorProfile.KubernetesConfig.CloudProviderBackoffRetries))
-		addValue(parametersMap, "cloudProviderBackoffExponent", strconv.FormatFloat(properties.OrchestratorProfile.KubernetesConfig.CloudProviderBackoffExponent, 'f', -1, 64))
-		addValue(parametersMap, "cloudProviderBackoffDuration", strconv.Itoa(properties.OrchestratorProfile.KubernetesConfig.CloudProviderBackoffDuration))
-		addValue(parametersMap, "cloudProviderBackoffJitter", strconv.FormatFloat(properties.OrchestratorProfile.KubernetesConfig.CloudProviderBackoffJitter, 'f', -1, 64))
-		addValue(parametersMap, "cloudProviderRatelimit", strconv.FormatBool(properties.OrchestratorProfile.KubernetesConfig.CloudProviderRateLimit))
-		addValue(parametersMap, "cloudProviderRatelimitQPS", strconv.FormatFloat(properties.OrchestratorProfile.KubernetesConfig.CloudProviderRateLimitQPS, 'f', -1, 64))
-		addValue(parametersMap, "cloudProviderRatelimitBucket", strconv.Itoa(properties.OrchestratorProfile.KubernetesConfig.CloudProviderRateLimitBucket))
-		addValue(parametersMap, "kubeClusterCidr", properties.OrchestratorProfile.KubernetesConfig.ClusterSubnet)
-		addValue(parametersMap, "kubernetesNonMasqueradeCidr", properties.OrchestratorProfile.KubernetesConfig.KubeletConfig["--non-masquerade-cidr"])
-		addValue(parametersMap, "kubernetesKubeletClusterDomain", properties.OrchestratorProfile.KubernetesConfig.KubeletConfig["--cluster-domain"])
-		addValue(parametersMap, "generatorCode", generatorCode)
-		if properties.HostedMasterProfile != nil {
-			addValue(parametersMap, "orchestratorName", "aks")
-		} else if properties.OrchestratorProfile.IsOpenShift() {
-			addValue(parametersMap, "orchestratorName", DefaultOpenshiftOrchestratorName)
-		} else {
-			addValue(parametersMap, "orchestratorName", DefaultOrchestratorName)
-		}
-		addValue(parametersMap, "dockerBridgeCidr", properties.OrchestratorProfile.KubernetesConfig.DockerBridgeSubnet)
-		addValue(parametersMap, "networkPolicy", properties.OrchestratorProfile.KubernetesConfig.NetworkPolicy)
-		addValue(parametersMap, "networkPlugin", properties.OrchestratorProfile.KubernetesConfig.NetworkPlugin)
-		addValue(parametersMap, "containerRuntime", properties.OrchestratorProfile.KubernetesConfig.ContainerRuntime)
-		addValue(parametersMap, "cniPluginsURL", cloudSpecConfig.KubernetesSpecConfig.CNIPluginsDownloadURL)
-		addValue(parametersMap, "vnetCniLinuxPluginsURL", cloudSpecConfig.KubernetesSpecConfig.VnetCNILinuxPluginsDownloadURL)
-		addValue(parametersMap, "vnetCniWindowsPluginsURL", cloudSpecConfig.KubernetesSpecConfig.VnetCNIWindowsPluginsDownloadURL)
-		addValue(parametersMap, "gchighthreshold", properties.OrchestratorProfile.KubernetesConfig.GCHighThreshold)
-		addValue(parametersMap, "gclowthreshold", properties.OrchestratorProfile.KubernetesConfig.GCLowThreshold)
-		addValue(parametersMap, "etcdDownloadURLBase", cloudSpecConfig.KubernetesSpecConfig.EtcdDownloadURLBase)
-		addValue(parametersMap, "etcdVersion", cs.Properties.OrchestratorProfile.KubernetesConfig.EtcdVersion)
-		addValue(parametersMap, "etcdDiskSizeGB", cs.Properties.OrchestratorProfile.KubernetesConfig.EtcdDiskSizeGB)
-		addValue(parametersMap, "etcdEncryptionKey", cs.Properties.OrchestratorProfile.KubernetesConfig.EtcdEncryptionKey)
-		if cs.Properties.OrchestratorProfile.KubernetesConfig.PrivateJumpboxProvision() {
-			addValue(parametersMap, "jumpboxVMName", cs.Properties.OrchestratorProfile.KubernetesConfig.PrivateCluster.JumpboxProfile.Name)
-			addValue(parametersMap, "jumpboxVMSize", cs.Properties.OrchestratorProfile.KubernetesConfig.PrivateCluster.JumpboxProfile.VMSize)
-			addValue(parametersMap, "jumpboxUsername", cs.Properties.OrchestratorProfile.KubernetesConfig.PrivateCluster.JumpboxProfile.Username)
-			addValue(parametersMap, "jumpboxOSDiskSizeGB", cs.Properties.OrchestratorProfile.KubernetesConfig.PrivateCluster.JumpboxProfile.OSDiskSizeGB)
-			addValue(parametersMap, "jumpboxPublicKey", cs.Properties.OrchestratorProfile.KubernetesConfig.PrivateCluster.JumpboxProfile.PublicKey)
-			addValue(parametersMap, "jumpboxStorageProfile", cs.Properties.OrchestratorProfile.KubernetesConfig.PrivateCluster.JumpboxProfile.StorageProfile)
-		}
-
-		if properties.OrchestratorProfile.KubernetesConfig == nil ||
-			!properties.OrchestratorProfile.KubernetesConfig.UseManagedIdentity {
-
-			addValue(parametersMap, "servicePrincipalClientId", properties.ServicePrincipalProfile.ClientID)
-			if properties.ServicePrincipalProfile.KeyvaultSecretRef != nil {
-				addKeyvaultReference(parametersMap, "servicePrincipalClientSecret",
-					properties.ServicePrincipalProfile.KeyvaultSecretRef.VaultID,
-					properties.ServicePrincipalProfile.KeyvaultSecretRef.SecretName,
-					properties.ServicePrincipalProfile.KeyvaultSecretRef.SecretVersion)
-			} else {
-				addValue(parametersMap, "servicePrincipalClientSecret", properties.ServicePrincipalProfile.Secret)
-			}
-
-			if properties.OrchestratorProfile.KubernetesConfig != nil && helpers.IsTrueBoolPointer(properties.OrchestratorProfile.KubernetesConfig.EnableEncryptionWithExternalKms) && !properties.OrchestratorProfile.KubernetesConfig.UseManagedIdentity && properties.ServicePrincipalProfile.ObjectID != "" {
-				addValue(parametersMap, "servicePrincipalObjectId", properties.ServicePrincipalProfile.ObjectID)
-			}
-		}
-
-		if properties.AADProfile != nil {
-			addValue(parametersMap, "aadTenantId", properties.AADProfile.TenantID)
-			if properties.AADProfile.AdminGroupID != "" {
-				addValue(parametersMap, "aadAdminGroupId", properties.AADProfile.AdminGroupID)
-			}
-		}
-	}
-
-	if strings.HasPrefix(properties.OrchestratorProfile.OrchestratorType, api.DCOS) {
-		dcosBootstrapURL := cloudSpecConfig.DCOSSpecConfig.DCOS188BootstrapDownloadURL
-		dcosWindowsBootstrapURL := cloudSpecConfig.DCOSSpecConfig.DCOSWindowsBootstrapDownloadURL
-		dcosRepositoryURL := cloudSpecConfig.DCOSSpecConfig.DcosRepositoryURL
-		dcosClusterPackageListID := cloudSpecConfig.DCOSSpecConfig.DcosClusterPackageListID
-		dcosProviderPackageID := cloudSpecConfig.DCOSSpecConfig.DcosProviderPackageID
-
-		switch properties.OrchestratorProfile.OrchestratorType {
-		case api.DCOS:
-			switch properties.OrchestratorProfile.OrchestratorVersion {
-			case common.DCOSVersion1Dot8Dot8:
-				dcosBootstrapURL = cloudSpecConfig.DCOSSpecConfig.DCOS188BootstrapDownloadURL
-			case common.DCOSVersion1Dot9Dot0:
-				dcosBootstrapURL = cloudSpecConfig.DCOSSpecConfig.DCOS190BootstrapDownloadURL
-			case common.DCOSVersion1Dot9Dot8:
-				dcosBootstrapURL = cloudSpecConfig.DCOSSpecConfig.DCOS198BootstrapDownloadURL
-			case common.DCOSVersion1Dot10Dot0:
-				dcosBootstrapURL = cloudSpecConfig.DCOSSpecConfig.DCOS110BootstrapDownloadURL
-			default:
-				dcosBootstrapURL = getDCOSDefaultBootstrapInstallerURL(properties.OrchestratorProfile)
-			}
-		}
-
-		if properties.OrchestratorProfile.DcosConfig != nil {
-			if properties.OrchestratorProfile.DcosConfig.DcosWindowsBootstrapURL != "" {
-				dcosWindowsBootstrapURL = properties.OrchestratorProfile.DcosConfig.DcosWindowsBootstrapURL
-			}
-			if properties.OrchestratorProfile.DcosConfig.DcosBootstrapURL != "" {
-				dcosBootstrapURL = properties.OrchestratorProfile.DcosConfig.DcosBootstrapURL
-			}
-			if len(properties.OrchestratorProfile.DcosConfig.Registry) > 0 {
-				addValue(parametersMap, "registry", properties.OrchestratorProfile.DcosConfig.Registry)
-				addValue(parametersMap, "registryKey", base64.StdEncoding.EncodeToString([]byte(fmt.Sprintf("%s:%s", properties.OrchestratorProfile.DcosConfig.RegistryUser, properties.OrchestratorProfile.DcosConfig.RegistryPass))))
-			}
-			if properties.OrchestratorProfile.DcosConfig.DcosRepositoryURL != "" {
-				dcosRepositoryURL = properties.OrchestratorProfile.DcosConfig.DcosRepositoryURL
-			} else {
-				dcosRepositoryURL = getDCOSDefaultRepositoryURL(
-					properties.OrchestratorProfile.OrchestratorType,
-					properties.OrchestratorProfile.OrchestratorVersion)
-			}
-
-			if properties.OrchestratorProfile.DcosConfig.DcosClusterPackageListID != "" {
-				dcosClusterPackageListID = properties.OrchestratorProfile.DcosConfig.DcosClusterPackageListID
-			}
-
-			if properties.OrchestratorProfile.DcosConfig.DcosProviderPackageID != "" {
-				dcosProviderPackageID = properties.OrchestratorProfile.DcosConfig.DcosProviderPackageID
-			} else {
-				dcosProviderPackageID = getDCOSDefaultProviderPackageGUID(
-					properties.OrchestratorProfile.OrchestratorType,
-					properties.OrchestratorProfile.OrchestratorVersion,
-					properties.MasterProfile.Count)
-			}
-		}
-
-		addValue(parametersMap, "dcosBootstrapURL", dcosBootstrapURL)
-		addValue(parametersMap, "dcosWindowsBootstrapURL", dcosWindowsBootstrapURL)
-		addValue(parametersMap, "dcosRepositoryURL", dcosRepositoryURL)
-		addValue(parametersMap, "dcosClusterPackageListID", dcosClusterPackageListID)
-		addValue(parametersMap, "dcosProviderPackageID", dcosProviderPackageID)
-
-		if properties.OrchestratorProfile.DcosConfig.BootstrapProfile != nil {
-			addValue(parametersMap, "bootstrapStaticIP", properties.OrchestratorProfile.DcosConfig.BootstrapProfile.StaticIP)
-			addValue(parametersMap, "bootstrapVMSize", properties.OrchestratorProfile.DcosConfig.BootstrapProfile.VMSize)
-		}
-	}
-
-	// Agent parameters
-	for _, agentProfile := range properties.AgentPoolProfiles {
-		addValue(parametersMap, fmt.Sprintf("%sCount", agentProfile.Name), agentProfile.Count)
-		addValue(parametersMap, fmt.Sprintf("%sVMSize", agentProfile.Name), agentProfile.VMSize)
-		if agentProfile.IsCustomVNET() {
-			addValue(parametersMap, fmt.Sprintf("%sVnetSubnetID", agentProfile.Name), agentProfile.VnetSubnetID)
-		} else {
-			addValue(parametersMap, fmt.Sprintf("%sSubnet", agentProfile.Name), agentProfile.Subnet)
-		}
-		if len(agentProfile.Ports) > 0 {
-			addValue(parametersMap, fmt.Sprintf("%sEndpointDNSNamePrefix", agentProfile.Name), agentProfile.DNSPrefix)
-		}
-
-		// Unless distro is defined, default distro is configured by defaults#setAgentNetworkDefaults
-		//   Ignores Windows OS
-		if !(agentProfile.OSType == api.Windows) {
-			if agentProfile.ImageRef != nil {
-				addValue(parametersMap, fmt.Sprintf("%sosImageName", agentProfile.Name), agentProfile.ImageRef.Name)
-				addValue(parametersMap, fmt.Sprintf("%sosImageResourceGroup", agentProfile.Name), agentProfile.ImageRef.ResourceGroup)
-			}
-			addValue(parametersMap, fmt.Sprintf("%sosImageOffer", agentProfile.Name), cloudSpecConfig.OSImageConfig[agentProfile.Distro].ImageOffer)
-			addValue(parametersMap, fmt.Sprintf("%sosImageSKU", agentProfile.Name), cloudSpecConfig.OSImageConfig[agentProfile.Distro].ImageSku)
-			addValue(parametersMap, fmt.Sprintf("%sosImagePublisher", agentProfile.Name), cloudSpecConfig.OSImageConfig[agentProfile.Distro].ImagePublisher)
-			addValue(parametersMap, fmt.Sprintf("%sosImageVersion", agentProfile.Name), cloudSpecConfig.OSImageConfig[agentProfile.Distro].ImageVersion)
-		}
-	}
-
-	// Windows parameters
-	if properties.HasWindows() {
-		addValue(parametersMap, "windowsAdminUsername", properties.WindowsProfile.AdminUsername)
-		addSecret(parametersMap, "windowsAdminPassword", properties.WindowsProfile.AdminPassword, false)
-		if properties.WindowsProfile.ImageVersion != "" {
-			addValue(parametersMap, "agentWindowsVersion", properties.WindowsProfile.ImageVersion)
-		}
-		if properties.WindowsProfile.WindowsImageSourceURL != "" {
-			addValue(parametersMap, "agentWindowsSourceUrl", properties.WindowsProfile.WindowsImageSourceURL)
-		}
-		if properties.WindowsProfile.WindowsPublisher != "" {
-			addValue(parametersMap, "agentWindowsPublisher", properties.WindowsProfile.WindowsPublisher)
-		}
-		if properties.WindowsProfile.WindowsOffer != "" {
-			addValue(parametersMap, "agentWindowsOffer", properties.WindowsProfile.WindowsOffer)
-		}
-		if properties.WindowsProfile.WindowsSku != "" {
-			addValue(parametersMap, "agentWindowsSku", properties.WindowsProfile.WindowsSku)
-		}
-		if properties.OrchestratorProfile.IsKubernetes() || properties.OrchestratorProfile.IsOpenShift() {
-			k8sVersion := properties.OrchestratorProfile.OrchestratorVersion
-			kubeBinariesSASURL := properties.OrchestratorProfile.KubernetesConfig.CustomWindowsPackageURL
-			if kubeBinariesSASURL == "" {
-				kubeBinariesSASURL = cloudSpecConfig.KubernetesSpecConfig.KubeBinariesSASURLBase + KubeConfigs[k8sVersion]["windowszip"]
-			}
-
-			addValue(parametersMap, "kubeBinariesSASURL", kubeBinariesSASURL)
-			addValue(parametersMap, "windowsPackageSASURLBase", cloudSpecConfig.KubernetesSpecConfig.WindowsPackageSASURLBase)
-			addValue(parametersMap, "kubeBinariesVersion", k8sVersion)
-			addValue(parametersMap, "windowsTelemetryGUID", cloudSpecConfig.KubernetesSpecConfig.WindowsTelemetryGUID)
-		}
-		for i, s := range properties.WindowsProfile.Secrets {
-			addValue(parametersMap, fmt.Sprintf("windowsKeyVaultID%d", i), s.SourceVault.ID)
-			for j, c := range s.VaultCertificates {
-				addValue(parametersMap, fmt.Sprintf("windowsKeyVaultID%dCertificateURL%d", i, j), c.CertificateURL)
-				addValue(parametersMap, fmt.Sprintf("windowsKeyVaultID%dCertificateStore%d", i, j), c.CertificateStore)
-			}
-		}
-	}
-
-	for _, extension := range properties.ExtensionProfiles {
-		if extension.ExtensionParametersKeyVaultRef != nil {
-			addKeyvaultReference(parametersMap, fmt.Sprintf("%sParameters", extension.Name),
-				extension.ExtensionParametersKeyVaultRef.VaultID,
-				extension.ExtensionParametersKeyVaultRef.SecretName,
-				extension.ExtensionParametersKeyVaultRef.SecretVersion)
-		} else {
-			addValue(parametersMap, fmt.Sprintf("%sParameters", extension.Name), extension.ExtensionParameters)
-		}
-	}
-
-	return parametersMap, nil
+	return ret
 }
 
 func addValue(m paramsMap, k string, v interface{}) {
@@ -1017,988 +268,6 @@ func getStorageAccountType(sizeName string) (string, error) {
 		return "Premium_LRS", nil
 	}
 	return "Standard_LRS", nil
-}
-
-// getTemplateFuncMap returns all functions used in template generation
-func (t *TemplateGenerator) getTemplateFuncMap(cs *api.ContainerService) template.FuncMap {
-	return template.FuncMap{
-		"IsHostedMaster": func() bool {
-			return cs.Properties.HostedMasterProfile != nil
-		},
-		"IsDCOS19": func() bool {
-			return cs.Properties.OrchestratorProfile.OrchestratorType == api.DCOS &&
-				(cs.Properties.OrchestratorProfile.OrchestratorVersion == common.DCOSVersion1Dot9Dot0 ||
-					cs.Properties.OrchestratorProfile.OrchestratorVersion == common.DCOSVersion1Dot9Dot8)
-		},
-		"IsKubernetesVersionGe": func(version string) bool {
-			return cs.Properties.OrchestratorProfile.IsKubernetes() && common.IsKubernetesVersionGe(cs.Properties.OrchestratorProfile.OrchestratorVersion, version)
-		},
-		"IsKubernetesVersionLt": func(version string) bool {
-			orchestratorVersion, _ := semver.NewVersion(cs.Properties.OrchestratorProfile.OrchestratorVersion)
-			constraint, _ := semver.NewConstraint("<" + version)
-			return cs.Properties.OrchestratorProfile.IsKubernetes() && constraint.Check(orchestratorVersion)
-		},
-		"IsKubernetesVersionTilde": func(version string) bool {
-			// examples include
-			// ~2.3 is equivalent to >= 2.3, < 2.4
-			// ~1.2.x is equivalent to >= 1.2.0, < 1.3.0
-			orchestratorVersion, _ := semver.NewVersion(cs.Properties.OrchestratorProfile.OrchestratorVersion)
-			constraint, _ := semver.NewConstraint("~" + version)
-			return cs.Properties.OrchestratorProfile.IsKubernetes() && constraint.Check(orchestratorVersion)
-		},
-		"GetMasterKubernetesLabels": func(rg string) string {
-			var buf bytes.Buffer
-			buf.WriteString("kubernetes.io/role=master")
-			buf.WriteString(fmt.Sprintf(",kubernetes.azure.com/cluster=%s", rg))
-			return buf.String()
-		},
-		"GetAgentKubernetesLabels": func(profile *api.AgentPoolProfile, rg string) string {
-			var buf bytes.Buffer
-			buf.WriteString(fmt.Sprintf("kubernetes.io/role=agent,agentpool=%s", profile.Name))
-			if profile.StorageProfile == api.ManagedDisks {
-				storagetier, _ := getStorageAccountType(profile.VMSize)
-				buf.WriteString(fmt.Sprintf(",storageprofile=managed,storagetier=%s", storagetier))
-			}
-			buf.WriteString(fmt.Sprintf(",kubernetes.azure.com/cluster=%s", rg))
-			for k, v := range profile.CustomNodeLabels {
-				buf.WriteString(fmt.Sprintf(",%s=%s", k, v))
-			}
-			return buf.String()
-		},
-		"GetKubeletConfigKeyVals": func(kc *api.KubernetesConfig) string {
-			if kc == nil {
-				return ""
-			}
-			kubeletConfig := cs.Properties.OrchestratorProfile.KubernetesConfig.KubeletConfig
-			if kc.KubeletConfig != nil {
-				kubeletConfig = kc.KubeletConfig
-			}
-			// Order by key for consistency
-			keys := []string{}
-			for key := range kubeletConfig {
-				keys = append(keys, key)
-			}
-			sort.Strings(keys)
-			var buf bytes.Buffer
-			for _, key := range keys {
-				buf.WriteString(fmt.Sprintf("%s=%s ", key, kubeletConfig[key]))
-			}
-			return buf.String()
-		},
-		"GetK8sRuntimeConfigKeyVals": func(config map[string]string) string {
-			// Order by key for consistency
-			keys := []string{}
-			for key := range config {
-				keys = append(keys, key)
-			}
-			sort.Strings(keys)
-			var buf bytes.Buffer
-			for _, key := range keys {
-				buf.WriteString(fmt.Sprintf("\\\"%s=%s\\\", ", key, config[key]))
-			}
-			return strings.TrimSuffix(buf.String(), ", ")
-		},
-		"HasPrivateRegistry": func() bool {
-			if cs.Properties.OrchestratorProfile.DcosConfig != nil {
-				return len(cs.Properties.OrchestratorProfile.DcosConfig.Registry) > 0
-			}
-			return false
-		},
-		"RequiresFakeAgentOutput": func() bool {
-			return cs.Properties.OrchestratorProfile.IsKubernetes() || cs.Properties.OrchestratorProfile.IsOpenShift()
-		},
-		"IsSwarmMode": func() bool {
-			return cs.Properties.OrchestratorProfile.IsSwarmMode()
-		},
-		"IsKubernetes": func() bool {
-			return cs.Properties.OrchestratorProfile.IsKubernetes()
-		},
-		"IsOpenShift": func() bool {
-			return cs.Properties.OrchestratorProfile.IsOpenShift()
-		},
-		"IsPublic": func(ports []int) bool {
-			return len(ports) > 0
-		},
-		"IsAzureCNI": func() bool {
-			return cs.Properties.OrchestratorProfile.IsAzureCNI()
-		},
-		"RequireRouteTable": func() bool {
-			return cs.Properties.OrchestratorProfile.RequireRouteTable()
-		},
-
-		"IsPrivateCluster": func() bool {
-			if !cs.Properties.OrchestratorProfile.IsKubernetes() {
-				return false
-			}
-			return helpers.IsTrueBoolPointer(cs.Properties.OrchestratorProfile.KubernetesConfig.PrivateCluster.Enabled)
-		},
-		"ProvisionJumpbox": func() bool {
-			return cs.Properties.OrchestratorProfile.KubernetesConfig.PrivateJumpboxProvision()
-		},
-		"JumpboxIsManagedDisks": func() bool {
-			if cs.Properties.OrchestratorProfile.KubernetesConfig.PrivateJumpboxProvision() && cs.Properties.OrchestratorProfile.KubernetesConfig.PrivateCluster.JumpboxProfile.StorageProfile == api.ManagedDisks {
-				return true
-			}
-			return false
-		},
-		"GetKubeConfig": func() string {
-			kubeConfig, err := GenerateKubeConfig(cs.Properties, cs.Location)
-			if err != nil {
-				return ""
-			}
-			return escapeSingleLine(kubeConfig)
-		},
-		"UseManagedIdentity": func() bool {
-			return cs.Properties.OrchestratorProfile.KubernetesConfig.UseManagedIdentity
-		},
-		"UseInstanceMetadata": func() bool {
-			return helpers.IsTrueBoolPointer(cs.Properties.OrchestratorProfile.KubernetesConfig.UseInstanceMetadata)
-		},
-		"GetVNETSubnetDependencies": func() string {
-			return getVNETSubnetDependencies(cs.Properties)
-		},
-		"GetLBRules": func(name string, ports []int) string {
-			return getLBRules(name, ports)
-		},
-		"GetProbes": func(ports []int) string {
-			return getProbes(ports)
-		},
-		"GetSecurityRules": func(ports []int) string {
-			return getSecurityRules(ports)
-		},
-		"GetUniqueNameSuffix": func() string {
-			return GenerateClusterID(cs.Properties)
-		},
-		"GetVNETAddressPrefixes": func() string {
-			return getVNETAddressPrefixes(cs.Properties)
-		},
-		"GetVNETSubnets": func(addNSG bool) string {
-			return getVNETSubnets(cs.Properties, addNSG)
-		},
-		"GetDataDisks": func(profile *api.AgentPoolProfile) string {
-			return getDataDisks(profile)
-		},
-		"HasBootstrap": func() bool {
-			return cs.Properties.OrchestratorProfile.DcosConfig != nil && cs.Properties.OrchestratorProfile.DcosConfig.BootstrapProfile != nil
-		},
-		"IsHostedBootstrap": func() bool {
-			return false
-		},
-		"GetDCOSBootstrapCustomData": func() string {
-			masterIPList := generateIPList(cs.Properties.MasterProfile.Count, cs.Properties.MasterProfile.FirstConsecutiveStaticIP)
-			for i, v := range masterIPList {
-				masterIPList[i] = "    - " + v
-			}
-
-			str := getSingleLineDCOSCustomData(
-				cs.Properties.OrchestratorProfile.OrchestratorType,
-				dcos2BootstrapCustomdata, 0,
-				map[string]string{
-					"PROVISION_SOURCE_STR":    getDCOSProvisionScript(dcosProvisionSource),
-					"PROVISION_STR":           getDCOSProvisionScript(dcos2BootstrapProvision),
-					"MASTER_IP_LIST":          strings.Join(masterIPList, "\n"),
-					"BOOTSTRAP_IP":            cs.Properties.OrchestratorProfile.DcosConfig.BootstrapProfile.StaticIP,
-					"BOOTSTRAP_OAUTH_ENABLED": strconv.FormatBool(cs.Properties.OrchestratorProfile.DcosConfig.BootstrapProfile.OAuthEnabled)})
-
-			return fmt.Sprintf("\"customData\": \"[base64(concat('#cloud-config\\n\\n', '%s'))]\",", str)
-		},
-		"GetDCOSMasterCustomData": func() string {
-			masterAttributeContents := getDCOSMasterCustomNodeLabels()
-			masterPreprovisionExtension := ""
-			if cs.Properties.MasterProfile.PreprovisionExtension != nil {
-				masterPreprovisionExtension += "\n"
-				masterPreprovisionExtension += makeMasterExtensionScriptCommands(cs)
-			}
-			var bootstrapIP string
-			if cs.Properties.OrchestratorProfile.DcosConfig != nil && cs.Properties.OrchestratorProfile.DcosConfig.BootstrapProfile != nil {
-				bootstrapIP = cs.Properties.OrchestratorProfile.DcosConfig.BootstrapProfile.StaticIP
-			}
-
-			str := getSingleLineDCOSCustomData(
-				cs.Properties.OrchestratorProfile.OrchestratorType,
-				getDCOSCustomDataTemplate(cs.Properties.OrchestratorProfile.OrchestratorType, cs.Properties.OrchestratorProfile.OrchestratorVersion),
-				cs.Properties.MasterProfile.Count,
-				map[string]string{
-					"PROVISION_SOURCE_STR":   getDCOSProvisionScript(dcosProvisionSource),
-					"PROVISION_STR":          getDCOSMasterProvisionScript(cs.Properties.OrchestratorProfile, bootstrapIP),
-					"ATTRIBUTES_STR":         masterAttributeContents,
-					"PREPROVISION_EXTENSION": masterPreprovisionExtension,
-					"ROLENAME":               "master"})
-
-			return fmt.Sprintf("\"customData\": \"[base64(concat('#cloud-config\\n\\n', '%s'))]\",", str)
-		},
-		"GetDCOSAgentCustomData": func(profile *api.AgentPoolProfile) string {
-			attributeContents := getDCOSAgentCustomNodeLabels(profile)
-			agentPreprovisionExtension := ""
-			if profile.PreprovisionExtension != nil {
-				agentPreprovisionExtension += "\n"
-				agentPreprovisionExtension += makeAgentExtensionScriptCommands(cs, profile)
-			}
-			var agentRoleName, bootstrapIP string
-			if len(profile.Ports) > 0 {
-				agentRoleName = "slave_public"
-			} else {
-				agentRoleName = "slave"
-			}
-			if cs.Properties.OrchestratorProfile.DcosConfig != nil && cs.Properties.OrchestratorProfile.DcosConfig.BootstrapProfile != nil {
-				bootstrapIP = cs.Properties.OrchestratorProfile.DcosConfig.BootstrapProfile.StaticIP
-			}
-
-			str := getSingleLineDCOSCustomData(
-				cs.Properties.OrchestratorProfile.OrchestratorType,
-				getDCOSCustomDataTemplate(cs.Properties.OrchestratorProfile.OrchestratorType, cs.Properties.OrchestratorProfile.OrchestratorVersion),
-				cs.Properties.MasterProfile.Count,
-				map[string]string{
-					"PROVISION_SOURCE_STR":   getDCOSProvisionScript(dcosProvisionSource),
-					"PROVISION_STR":          getDCOSAgentProvisionScript(profile, cs.Properties.OrchestratorProfile, bootstrapIP),
-					"ATTRIBUTES_STR":         attributeContents,
-					"PREPROVISION_EXTENSION": agentPreprovisionExtension,
-					"ROLENAME":               agentRoleName})
-
-			return fmt.Sprintf("\"customData\": \"[base64(concat('#cloud-config\\n\\n', '%s'))]\",", str)
-		},
-		"GetDCOSWindowsAgentCustomData": func(profile *api.AgentPoolProfile) string {
-			agentPreprovisionExtension := ""
-			if profile.PreprovisionExtension != nil {
-				agentPreprovisionExtension += "\n"
-				agentPreprovisionExtension += makeAgentExtensionScriptCommands(cs, profile)
-			}
-			b, err := Asset(dcosWindowsProvision)
-			if err != nil {
-				// this should never happen and this is a bug
-				panic(fmt.Sprintf("BUG: %s", err.Error()))
-			}
-			// translate the parameters
-			csStr := string(b)
-			csStr = strings.Replace(csStr, "PREPROVISION_EXTENSION", agentPreprovisionExtension, -1)
-			csStr = strings.Replace(csStr, "\r\n", "\n", -1)
-			str := getBase64CustomScriptFromStr(csStr)
-			return fmt.Sprintf("\"customData\": \"%s\"", str)
-		},
-		"GetDCOSWindowsAgentCustomNodeAttributes": func(profile *api.AgentPoolProfile) string {
-			return getDCOSWindowsAgentCustomAttributes(profile)
-		},
-		"GetDCOSWindowsAgentPreprovisionParameters": func(profile *api.AgentPoolProfile) string {
-			agentPreprovisionExtensionParameters := ""
-			if profile.PreprovisionExtension != nil {
-				agentPreprovisionExtensionParameters = getDCOSWindowsAgentPreprovisionParameters(cs, profile)
-			}
-			return agentPreprovisionExtensionParameters
-		},
-		"GetMasterAllowedSizes": func() string {
-			if t.ClassicMode {
-				return GetClassicAllowedSizes()
-			} else if cs.Properties.OrchestratorProfile.OrchestratorType == api.DCOS {
-				return GetDCOSMasterAllowedSizes()
-			}
-			return GetMasterAgentAllowedSizes()
-		},
-		"GetAgentAllowedSizes": func() string {
-			if t.ClassicMode {
-				return GetClassicAllowedSizes()
-			} else if cs.Properties.OrchestratorProfile.IsKubernetes() || cs.Properties.OrchestratorProfile.IsOpenShift() {
-				return GetKubernetesAgentAllowedSizes()
-			}
-			return GetMasterAgentAllowedSizes()
-		},
-		"getSwarmVersions": func() string {
-			return getSwarmVersions(api.SwarmVersion, api.SwarmDockerComposeVersion)
-		},
-		"GetSwarmModeVersions": func() string {
-			return getSwarmVersions(api.DockerCEVersion, api.DockerCEDockerComposeVersion)
-		},
-		"GetSizeMap": func() string {
-			if t.ClassicMode {
-				return GetClassicSizeMap()
-			}
-			return GetSizeMap()
-		},
-		"GetClassicMode": func() bool {
-			return t.ClassicMode
-		},
-		"Base64": func(s string) string {
-			return base64.StdEncoding.EncodeToString([]byte(s))
-		},
-		"GetDefaultInternalLbStaticIPOffset": func() int {
-			return DefaultInternalLbStaticIPOffset
-		},
-		"GetKubernetesMasterCustomData": func(profile *api.Properties) string {
-			str, e := t.getSingleLineForTemplate(kubernetesMasterCustomDataYaml, cs, profile)
-			if e != nil {
-				fmt.Printf("%#v\n", e)
-				return ""
-			}
-
-			// add manifests
-			str = substituteConfigString(str,
-				kubernetesManifestSettingsInit(profile),
-				"k8s/manifests",
-				"/etc/kubernetes/manifests",
-				"MASTER_MANIFESTS_CONFIG_PLACEHOLDER",
-				profile.OrchestratorProfile.OrchestratorVersion)
-
-			// add artifacts
-			str = substituteConfigString(str,
-				kubernetesArtifactSettingsInitMaster(profile),
-				"k8s/artifacts",
-				"/etc/systemd/system",
-				"MASTER_ARTIFACTS_CONFIG_PLACEHOLDER",
-				profile.OrchestratorProfile.OrchestratorVersion)
-
-			// add addons
-			str = substituteConfigString(str,
-				kubernetesAddonSettingsInit(profile),
-				"k8s/addons",
-				"/etc/kubernetes/addons",
-				"MASTER_ADDONS_CONFIG_PLACEHOLDER",
-				profile.OrchestratorProfile.OrchestratorVersion)
-
-			// return the custom data
-			return fmt.Sprintf("\"customData\": \"[base64(concat('%s'))]\",", str)
-		},
-		"GetKubernetesAgentCustomData": func(profile *api.AgentPoolProfile) string {
-			str, e := t.getSingleLineForTemplate(kubernetesAgentCustomDataYaml, cs, profile)
-
-			if e != nil {
-				return ""
-			}
-
-			// add artifacts
-			str = substituteConfigString(str,
-				kubernetesArtifactSettingsInitAgent(cs.Properties),
-				"k8s/artifacts",
-				"/etc/systemd/system",
-				"AGENT_ARTIFACTS_CONFIG_PLACEHOLDER",
-				cs.Properties.OrchestratorProfile.OrchestratorVersion)
-
-			return fmt.Sprintf("\"customData\": \"[base64(concat('%s'))]\",", str)
-		},
-		"GetKubernetesJumpboxCustomData": func(p *api.Properties) string {
-			str, err := t.getSingleLineForTemplate(kubernetesJumpboxCustomDataYaml, cs, p)
-
-			if err != nil {
-				return ""
-			}
-
-			return fmt.Sprintf("\"customData\": \"[base64(concat('%s'))]\",", str)
-		},
-		"WriteLinkedTemplatesForExtensions": func() string {
-			extensions := getLinkedTemplatesForExtensions(cs.Properties)
-			return extensions
-		},
-		"GetKubernetesB64Provision": func() string {
-			return getBase64CustomScript(kubernetesCustomScript)
-		},
-		"GetKubernetesB64ProvisionSource": func() string {
-			return getBase64CustomScript(kubernetesProvisionSourceScript)
-		},
-		"GetKubernetesB64Mountetcd": func() string {
-			return getBase64CustomScript(kubernetesMountetcd)
-		},
-		"GetKubernetesB64CustomSearchDomainsScript": func() string {
-			return getBase64CustomScript(kubernetesCustomSearchDomainsScript)
-		},
-		"GetKubernetesB64GenerateProxyCerts": func() string {
-			return getBase64CustomScript(kubernetesMasterGenerateProxyCertsScript)
-		},
-		"GetKubernetesMasterPreprovisionYaml": func() string {
-			str := ""
-			if cs.Properties.MasterProfile.PreprovisionExtension != nil {
-				str += "\n"
-				str += makeMasterExtensionScriptCommands(cs)
-			}
-			return str
-		},
-		"GetKubernetesAgentPreprovisionYaml": func(profile *api.AgentPoolProfile) string {
-			str := ""
-			if profile.PreprovisionExtension != nil {
-				str += "\n"
-				str += makeAgentExtensionScriptCommands(cs, profile)
-			}
-			return str
-		},
-		"GetMasterSwarmCustomData": func() string {
-			files := []string{swarmProvision}
-			str := buildYamlFileWithWriteFiles(files)
-			if cs.Properties.MasterProfile.PreprovisionExtension != nil {
-				extensionStr := makeMasterExtensionScriptCommands(cs)
-				str += "'runcmd:\n" + extensionStr + "\n\n'"
-			}
-			str = escapeSingleLine(str)
-			return fmt.Sprintf("\"customData\": \"[base64(concat('%s'))]\",", str)
-		},
-		"GetAgentSwarmCustomData": func(profile *api.AgentPoolProfile) string {
-			files := []string{swarmProvision}
-			str := buildYamlFileWithWriteFiles(files)
-			str = escapeSingleLine(str)
-			return fmt.Sprintf("\"customData\": \"[base64(concat('%s',variables('%sRunCmdFile'),variables('%sRunCmd')))]\",", str, profile.Name, profile.Name)
-		},
-		"GetSwarmAgentPreprovisionExtensionCommands": func(profile *api.AgentPoolProfile) string {
-			str := ""
-			if profile.PreprovisionExtension != nil {
-				makeAgentExtensionScriptCommands(cs, profile)
-			}
-			str = escapeSingleLine(str)
-			return str
-		},
-		"GetLocation": func() string {
-			return cs.Location
-		},
-		"GetWinAgentSwarmCustomData": func() string {
-			str := getBase64CustomScript(swarmWindowsProvision)
-			return fmt.Sprintf("\"customData\": \"%s\"", str)
-		},
-		"GetWinAgentSwarmModeCustomData": func() string {
-			str := getBase64CustomScript(swarmModeWindowsProvision)
-			return fmt.Sprintf("\"customData\": \"%s\"", str)
-		},
-		"GetKubernetesWindowsAgentCustomData": func(profile *api.AgentPoolProfile) string {
-			str, e := t.getSingleLineForTemplate(kubernetesWindowsAgentCustomDataPS1, cs, profile)
-			if e != nil {
-				return ""
-			}
-			return fmt.Sprintf("\"customData\": \"[base64(concat('%s'))]\",", str)
-		},
-		"GetMasterSwarmModeCustomData": func() string {
-			files := []string{swarmModeProvision}
-			str := buildYamlFileWithWriteFiles(files)
-			if cs.Properties.MasterProfile.PreprovisionExtension != nil {
-				extensionStr := makeMasterExtensionScriptCommands(cs)
-				str += "runcmd:\n" + extensionStr + "\n\n"
-			}
-			str = escapeSingleLine(str)
-			return fmt.Sprintf("\"customData\": \"[base64(concat('%s'))]\",", str)
-		},
-		"GetAgentSwarmModeCustomData": func(profile *api.AgentPoolProfile) string {
-			files := []string{swarmModeProvision}
-			str := buildYamlFileWithWriteFiles(files)
-			str = escapeSingleLine(str)
-			return fmt.Sprintf("\"customData\": \"[base64(concat('%s',variables('%sRunCmdFile'),variables('%sRunCmd')))]\",", str, profile.Name, profile.Name)
-		},
-		"GetKubernetesSubnets": func() string {
-			return getKubernetesSubnets(cs.Properties)
-		},
-		"GetKubernetesPodStartIndex": func() string {
-			return fmt.Sprintf("%d", getKubernetesPodStartIndex(cs.Properties))
-		},
-		"WrapAsVariable": func(s string) string {
-			return fmt.Sprintf("',variables('%s'),'", s)
-		},
-		"WrapAsVerbatim": func(s string) string {
-			return fmt.Sprintf("',%s,'", s)
-		},
-		"AnyAgentUsesAvailabilitySets": func() bool {
-			for _, agentProfile := range cs.Properties.AgentPoolProfiles {
-				if agentProfile.IsAvailabilitySets() {
-					return true
-				}
-			}
-			return false
-		},
-		"AnyAgentUsesVirtualMachineScaleSets": func() bool {
-			for _, agentProfile := range cs.Properties.AgentPoolProfiles {
-				if agentProfile.IsVirtualMachineScaleSets() {
-					return true
-				}
-			}
-			return false
-		},
-		"HasLinuxAgents": func() bool {
-			for _, agentProfile := range cs.Properties.AgentPoolProfiles {
-				if agentProfile.IsLinux() {
-					return true
-				}
-			}
-			return false
-		},
-		"IsNVIDIADevicePluginEnabled": func() bool {
-			return cs.Properties.IsNVIDIADevicePluginEnabled()
-		},
-		"IsNSeriesSKU": func(profile *api.AgentPoolProfile) bool {
-			return isNSeriesSKU(profile)
-		},
-		"GetGPUDriversInstallScript": func(profile *api.AgentPoolProfile) string {
-			return getGPUDriversInstallScript(profile)
-		},
-		"HasLinuxSecrets": func() bool {
-			return cs.Properties.LinuxProfile.HasSecrets()
-		},
-		"HasCustomSearchDomain": func() bool {
-			return cs.Properties.LinuxProfile.HasSearchDomain()
-		},
-		"HasWindowsSecrets": func() bool {
-			return cs.Properties.WindowsProfile.HasSecrets()
-		},
-		"HasWindowsCustomImage": func() bool {
-			return cs.Properties.WindowsProfile.HasCustomImage()
-		},
-		"GetConfigurationScriptRootURL": func() string {
-			if cs.Properties.LinuxProfile.ScriptRootURL == "" {
-				return DefaultConfigurationScriptRootURL
-			}
-			return cs.Properties.LinuxProfile.ScriptRootURL
-		},
-		"GetMasterOSImageOffer": func() string {
-			cloudSpecConfig := GetCloudSpecConfig(cs.Location)
-			return fmt.Sprintf("\"%s\"", cloudSpecConfig.OSImageConfig[cs.Properties.MasterProfile.Distro].ImageOffer)
-		},
-		"GetMasterOSImagePublisher": func() string {
-			cloudSpecConfig := GetCloudSpecConfig(cs.Location)
-			return fmt.Sprintf("\"%s\"", cloudSpecConfig.OSImageConfig[cs.Properties.MasterProfile.Distro].ImagePublisher)
-		},
-		"GetMasterOSImageSKU": func() string {
-			cloudSpecConfig := GetCloudSpecConfig(cs.Location)
-			return fmt.Sprintf("\"%s\"", cloudSpecConfig.OSImageConfig[cs.Properties.MasterProfile.Distro].ImageSku)
-		},
-		"GetMasterOSImageVersion": func() string {
-			cloudSpecConfig := GetCloudSpecConfig(cs.Location)
-			return fmt.Sprintf("\"%s\"", cloudSpecConfig.OSImageConfig[cs.Properties.MasterProfile.Distro].ImageVersion)
-		},
-		"GetAgentOSImageOffer": func(profile *api.AgentPoolProfile) string {
-			cloudSpecConfig := GetCloudSpecConfig(cs.Location)
-			return fmt.Sprintf("\"%s\"", cloudSpecConfig.OSImageConfig[profile.Distro].ImageOffer)
-		},
-		"GetAgentOSImagePublisher": func(profile *api.AgentPoolProfile) string {
-			cloudSpecConfig := GetCloudSpecConfig(cs.Location)
-			return fmt.Sprintf("\"%s\"", cloudSpecConfig.OSImageConfig[profile.Distro].ImagePublisher)
-		},
-		"GetAgentOSImageSKU": func(profile *api.AgentPoolProfile) string {
-			cloudSpecConfig := GetCloudSpecConfig(cs.Location)
-			return fmt.Sprintf("\"%s\"", cloudSpecConfig.OSImageConfig[profile.Distro].ImageSku)
-		},
-		"GetAgentOSImageVersion": func(profile *api.AgentPoolProfile) string {
-			cloudSpecConfig := GetCloudSpecConfig(cs.Location)
-			return fmt.Sprintf("\"%s\"", cloudSpecConfig.OSImageConfig[profile.Distro].ImageVersion)
-		},
-		"UseAgentCustomImage": func(profile *api.AgentPoolProfile) bool {
-			imageRef := profile.ImageRef
-			return imageRef != nil && len(imageRef.Name) > 0 && len(imageRef.ResourceGroup) > 0
-		},
-		"UseMasterCustomImage": func() bool {
-			imageRef := cs.Properties.MasterProfile.ImageRef
-			return imageRef != nil && len(imageRef.Name) > 0 && len(imageRef.ResourceGroup) > 0
-		},
-		"GetMasterEtcdServerPort": func() int {
-			return DefaultMasterEtcdServerPort
-		},
-		"GetMasterEtcdClientPort": func() int {
-			return DefaultMasterEtcdClientPort
-		},
-		"PopulateClassicModeDefaultValue": func(attr string) string {
-			var val string
-			if !t.ClassicMode {
-				val = ""
-			} else {
-				k8sVersion := cs.Properties.OrchestratorProfile.OrchestratorVersion
-				cloudSpecConfig := GetCloudSpecConfig(cs.Location)
-				tillerAddon := getAddonByName(cs.Properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultTillerAddonName)
-				tC := getAddonContainersIndexByName(tillerAddon.Containers, DefaultTillerAddonName)
-				aciConnectorAddon := getAddonByName(cs.Properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultACIConnectorAddonName)
-				aC := getAddonContainersIndexByName(aciConnectorAddon.Containers, DefaultACIConnectorAddonName)
-				clusterAutoscalerAddon := getAddonByName(cs.Properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultClusterAutoscalerAddonName)
-				aS := getAddonContainersIndexByName(clusterAutoscalerAddon.Containers, DefaultClusterAutoscalerAddonName)
-				dashboardAddon := getAddonByName(cs.Properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultDashboardAddonName)
-				dC := getAddonContainersIndexByName(dashboardAddon.Containers, DefaultDashboardAddonName)
-				reschedulerAddon := getAddonByName(cs.Properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultReschedulerAddonName)
-				rC := getAddonContainersIndexByName(reschedulerAddon.Containers, DefaultReschedulerAddonName)
-				metricsServerAddon := getAddonByName(cs.Properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultMetricsServerAddonName)
-				mC := getAddonContainersIndexByName(metricsServerAddon.Containers, DefaultMetricsServerAddonName)
-				nvidiaDevicePluginAddon := getAddonByName(cs.Properties.OrchestratorProfile.KubernetesConfig.Addons, DefaultNVIDIADevicePluginAddonName)
-				nC := getAddonContainersIndexByName(nvidiaDevicePluginAddon.Containers, DefaultNVIDIADevicePluginAddonName)
-				switch attr {
-				case "kubernetesHyperkubeSpec":
-					val = cs.Properties.OrchestratorProfile.KubernetesConfig.KubernetesImageBase + KubeConfigs[k8sVersion]["hyperkube"]
-					if cs.Properties.OrchestratorProfile.KubernetesConfig.CustomHyperkubeImage != "" {
-						val = cs.Properties.OrchestratorProfile.KubernetesConfig.CustomHyperkubeImage
-					}
-				case "dockerEngineVersion":
-					val = cs.Properties.OrchestratorProfile.KubernetesConfig.KubernetesImageBase + KubeConfigs[k8sVersion]["dockerEngineVersion"]
-					if cs.Properties.OrchestratorProfile.KubernetesConfig.DockerEngineVersion != "" {
-						val = cs.Properties.OrchestratorProfile.KubernetesConfig.DockerEngineVersion
-					}
-				case "kubernetesAddonManagerSpec":
-					val = cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase + KubeConfigs[k8sVersion]["addonmanager"]
-				case "kubernetesAddonResizerSpec":
-					val = cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase + KubeConfigs[k8sVersion]["addonresizer"]
-				case "kubernetesDashboardSpec":
-					if dC > -1 {
-						if dashboardAddon.Containers[dC].Image != "" {
-							val = dashboardAddon.Containers[dC].Image
-						}
-					} else {
-						val = cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase + KubeConfigs[k8sVersion][DefaultDashboardAddonName]
-					}
-				case "kubernetesDashboardCPURequests":
-					if dC > -1 {
-						val = dashboardAddon.Containers[dC].CPURequests
-					} else {
-						val = ""
-					}
-				case "kubernetesDashboardMemoryRequests":
-					if dC > -1 {
-						val = dashboardAddon.Containers[dC].MemoryRequests
-					} else {
-						val = ""
-					}
-				case "kubernetesDashboardCPULimit":
-					if dC > -1 {
-						val = dashboardAddon.Containers[dC].CPULimits
-					} else {
-						val = ""
-					}
-				case "kubernetesDashboardMemoryLimit":
-					if dC > -1 {
-						val = dashboardAddon.Containers[dC].MemoryLimits
-					} else {
-						val = ""
-					}
-				case "kubernetesDNSMasqSpec":
-					val = cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase + KubeConfigs[k8sVersion]["dnsmasq"]
-				case "kubernetesExecHealthzSpec":
-					val = cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase + KubeConfigs[k8sVersion]["exechealthz"]
-				case "kubernetesHeapsterSpec":
-					val = cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase + KubeConfigs[k8sVersion]["heapster"]
-				case "kubernetesACIConnectorSpec":
-					if aC > -1 {
-						if aciConnectorAddon.Containers[aC].Image != "" {
-							val = aciConnectorAddon.Containers[aC].Image
-						} else {
-							val = cloudSpecConfig.KubernetesSpecConfig.ACIConnectorImageBase + KubeConfigs[k8sVersion][DefaultACIConnectorAddonName]
-						}
-					}
-				case "kubernetesACIConnectorNodeName":
-					if aC > -1 {
-						val = aciConnectorAddon.Config["nodeName"]
-					} else {
-						val = ""
-					}
-				case "kubernetesACIConnectorOS":
-					if aC > -1 {
-						val = aciConnectorAddon.Config["os"]
-					} else {
-						val = ""
-					}
-				case "kubernetesACIConnectorTaint":
-					if aC > -1 {
-						val = aciConnectorAddon.Config["taint"]
-					} else {
-						val = ""
-					}
-				case "kubernetesACIConnectorRegion":
-					if aC > -1 {
-						val = aciConnectorAddon.Config["region"]
-					} else {
-						val = ""
-					}
-				case "kubernetesACIConnectorCPURequests":
-					if aC > -1 {
-						val = aciConnectorAddon.Containers[aC].CPURequests
-					} else {
-						val = ""
-					}
-				case "kubernetesACIConnectorMemoryRequests":
-					if aC > -1 {
-						val = aciConnectorAddon.Containers[aC].MemoryRequests
-					} else {
-						val = ""
-					}
-				case "kubernetesACIConnectorCPULimit":
-					if aC > -1 {
-						val = aciConnectorAddon.Containers[aC].CPULimits
-					} else {
-						val = ""
-					}
-				case "kubernetesACIConnectorMemoryLimit":
-					if aC > -1 {
-						val = aciConnectorAddon.Containers[aC].MemoryLimits
-					} else {
-						val = ""
-					}
-				case "kubernetesClusterAutoscalerSpec":
-					if aS > -1 {
-						if clusterAutoscalerAddon.Containers[aS].Image != "" {
-							val = clusterAutoscalerAddon.Containers[aS].Image
-						} else {
-							val = cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase + KubeConfigs[k8sVersion][DefaultClusterAutoscalerAddonName]
-						}
-					}
-				case "kubernetesClusterAutoscalerCPURequests":
-					if aS > -1 {
-						val = clusterAutoscalerAddon.Containers[aC].CPURequests
-					} else {
-						val = ""
-					}
-				case "kubernetesClusterAutoscalerMemoryRequests":
-					if aS > -1 {
-						val = clusterAutoscalerAddon.Containers[aC].MemoryRequests
-					} else {
-						val = ""
-					}
-				case "kubernetesClusterAutoscalerCPULimit":
-					if aS > -1 {
-						val = clusterAutoscalerAddon.Containers[aC].CPULimits
-					} else {
-						val = ""
-					}
-				case "kubernetesClusterAutoscalerMemoryLimit":
-					if aS > -1 {
-						val = clusterAutoscalerAddon.Containers[aC].MemoryLimits
-					} else {
-						val = ""
-					}
-				case "kubernetesClusterAutoscalerUseManagedIdentity":
-					if aS > -1 {
-						if cs.Properties.OrchestratorProfile.KubernetesConfig != nil && cs.Properties.OrchestratorProfile.KubernetesConfig.UseManagedIdentity {
-							val = strings.ToLower(strconv.FormatBool(cs.Properties.OrchestratorProfile.KubernetesConfig.UseManagedIdentity))
-						} else {
-							val = "false"
-						}
-					}
-				case "kubernetesTillerSpec":
-					if tC > -1 {
-						if tillerAddon.Containers[tC].Image != "" {
-							val = tillerAddon.Containers[tC].Image
-						} else {
-							val = cloudSpecConfig.KubernetesSpecConfig.TillerImageBase + KubeConfigs[k8sVersion][DefaultTillerAddonName]
-						}
-					}
-				case "kubernetesTillerCPURequests":
-					if tC > -1 {
-						val = tillerAddon.Containers[tC].CPURequests
-					} else {
-						val = ""
-					}
-				case "kubernetesTillerMemoryRequests":
-					if tC > -1 {
-						val = tillerAddon.Containers[tC].MemoryRequests
-					} else {
-						val = ""
-					}
-				case "kubernetesTillerCPULimit":
-					if tC > -1 {
-						val = tillerAddon.Containers[tC].CPULimits
-					} else {
-						val = ""
-					}
-				case "kubernetesTillerMemoryLimit":
-					if tC > -1 {
-						val = tillerAddon.Containers[tC].MemoryLimits
-					} else {
-						val = ""
-					}
-				case "kubernetesTillerMaxHistory":
-					if tC > -1 {
-						if _, ok := tillerAddon.Config["max-history"]; ok {
-							val = tillerAddon.Config["max-history"]
-						} else {
-							val = "0"
-						}
-					} else {
-						val = "0"
-					}
-				case "kubernetesMetricsServerSpec":
-					if mC > -1 {
-						if metricsServerAddon.Containers[mC].Image != "" {
-							val = metricsServerAddon.Containers[mC].Image
-						}
-					} else {
-						val = cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase + KubeConfigs[k8sVersion][DefaultMetricsServerAddonName]
-					}
-				case "kubernetesNVIDIADevicePluginSpec":
-					if nC > -1 {
-						if nvidiaDevicePluginAddon.Containers[nC].Image != "" {
-							val = nvidiaDevicePluginAddon.Containers[nC].Image
-						}
-					} else {
-						val = cloudSpecConfig.KubernetesSpecConfig.NVIDIAImageBase + KubeConfigs[k8sVersion][DefaultNVIDIADevicePluginAddonName]
-					}
-				case "kubernetesReschedulerSpec":
-					if rC > -1 {
-						if reschedulerAddon.Containers[rC].Image != "" {
-							val = reschedulerAddon.Containers[rC].Image
-						}
-					} else {
-						val = cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase + KubeConfigs[k8sVersion][DefaultReschedulerAddonName]
-					}
-				case "kubernetesReschedulerCPURequests":
-					if rC > -1 {
-						val = reschedulerAddon.Containers[rC].CPURequests
-					} else {
-						val = ""
-					}
-				case "kubernetesReschedulerMemoryRequests":
-					if rC > -1 {
-						val = reschedulerAddon.Containers[rC].MemoryRequests
-					} else {
-						val = ""
-					}
-				case "kubernetesReschedulerCPULimit":
-					if rC > -1 {
-						val = reschedulerAddon.Containers[rC].CPULimits
-					} else {
-						val = ""
-					}
-				case "kubernetesReschedulerMemoryLimit":
-					if rC > -1 {
-						val = reschedulerAddon.Containers[rC].MemoryLimits
-					} else {
-						val = ""
-					}
-				case "kubernetesKubeDNSSpec":
-					val = cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase + KubeConfigs[k8sVersion]["dns"]
-				case "kubernetesPodInfraContainerSpec":
-					val = cloudSpecConfig.KubernetesSpecConfig.KubernetesImageBase + KubeConfigs[k8sVersion]["pause"]
-				case "cloudProviderBackoff":
-					val = strconv.FormatBool(cs.Properties.OrchestratorProfile.KubernetesConfig.CloudProviderBackoff)
-				case "cloudProviderBackoffRetries":
-					val = strconv.Itoa(cs.Properties.OrchestratorProfile.KubernetesConfig.CloudProviderBackoffRetries)
-				case "cloudProviderBackoffExponent":
-					val = strconv.FormatFloat(cs.Properties.OrchestratorProfile.KubernetesConfig.CloudProviderBackoffExponent, 'f', -1, 64)
-				case "cloudProviderBackoffDuration":
-					val = strconv.Itoa(cs.Properties.OrchestratorProfile.KubernetesConfig.CloudProviderBackoffDuration)
-				case "cloudProviderBackoffJitter":
-					val = strconv.FormatFloat(cs.Properties.OrchestratorProfile.KubernetesConfig.CloudProviderBackoffJitter, 'f', -1, 64)
-				case "cloudProviderRatelimit":
-					val = strconv.FormatBool(cs.Properties.OrchestratorProfile.KubernetesConfig.CloudProviderRateLimit)
-				case "cloudProviderRatelimitQPS":
-					val = strconv.FormatFloat(cs.Properties.OrchestratorProfile.KubernetesConfig.CloudProviderRateLimitQPS, 'f', -1, 64)
-				case "cloudProviderRatelimitBucket":
-					val = strconv.Itoa(cs.Properties.OrchestratorProfile.KubernetesConfig.CloudProviderRateLimitBucket)
-				case "kubeBinariesSASURL":
-					val = cloudSpecConfig.KubernetesSpecConfig.KubeBinariesSASURLBase + KubeConfigs[k8sVersion]["windowszip"]
-				case "windowsPackageSASURLBase":
-					val = cloudSpecConfig.KubernetesSpecConfig.WindowsPackageSASURLBase
-				case "kubeClusterCidr":
-					val = DefaultKubernetesClusterSubnet
-				case "kubeDNSServiceIP":
-					val = DefaultKubernetesDNSServiceIP
-				case "kubeServiceCidr":
-					val = DefaultKubernetesServiceCIDR
-				case "kubeBinariesVersion":
-					val = cs.Properties.OrchestratorProfile.OrchestratorVersion
-				case "windowsTelemetryGUID":
-					val = cloudSpecConfig.KubernetesSpecConfig.WindowsTelemetryGUID
-				case "caPrivateKey":
-					// The base64 encoded "NotAvailable"
-					val = "Tm90QXZhaWxhYmxlCg=="
-				case "dockerBridgeCidr":
-					val = DefaultDockerBridgeSubnet
-				case "gchighthreshold":
-					val = strconv.Itoa(cs.Properties.OrchestratorProfile.KubernetesConfig.GCHighThreshold)
-				case "gclowthreshold":
-					val = strconv.Itoa(cs.Properties.OrchestratorProfile.KubernetesConfig.GCLowThreshold)
-				case "generatorCode":
-					val = DefaultGeneratorCode
-				case "orchestratorName":
-					val = DefaultOrchestratorName
-				case "etcdImageBase":
-					val = cloudSpecConfig.KubernetesSpecConfig.EtcdDownloadURLBase
-				case "etcdVersion":
-					val = cs.Properties.OrchestratorProfile.KubernetesConfig.EtcdVersion
-				case "etcdEncryptionKey":
-					val = cs.Properties.OrchestratorProfile.KubernetesConfig.EtcdEncryptionKey
-				case "etcdDiskSizeGB":
-					val = cs.Properties.OrchestratorProfile.KubernetesConfig.EtcdDiskSizeGB
-				case "jumpboxOSDiskSizeGB":
-					val = strconv.Itoa(cs.Properties.OrchestratorProfile.KubernetesConfig.PrivateCluster.JumpboxProfile.OSDiskSizeGB)
-				default:
-					val = ""
-				}
-			}
-			return fmt.Sprintf("\"defaultValue\": \"%s\",", val)
-		},
-		"UseCloudControllerManager": func() bool {
-			return cs.Properties.OrchestratorProfile.KubernetesConfig.UseCloudControllerManager != nil && *cs.Properties.OrchestratorProfile.KubernetesConfig.UseCloudControllerManager
-		},
-		"AdminGroupID": func() bool {
-			return cs.Properties.AADProfile != nil && cs.Properties.AADProfile.AdminGroupID != ""
-		},
-		"EnableDataEncryptionAtRest": func() bool {
-			return helpers.IsTrueBoolPointer(cs.Properties.OrchestratorProfile.KubernetesConfig.EnableDataEncryptionAtRest)
-		},
-		"EnableEncryptionWithExternalKms": func() bool {
-			return helpers.IsTrueBoolPointer(cs.Properties.OrchestratorProfile.KubernetesConfig.EnableEncryptionWithExternalKms)
-		},
-		"EnableAggregatedAPIs": func() bool {
-			if cs.Properties.OrchestratorProfile.KubernetesConfig.EnableAggregatedAPIs {
-				return true
-			} else if common.IsKubernetesVersionGe(cs.Properties.OrchestratorProfile.OrchestratorVersion, "1.9.0") {
-				return true
-			}
-			return false
-		},
-		"EnablePodSecurityPolicy": func() bool {
-			return helpers.IsTrueBoolPointer(cs.Properties.OrchestratorProfile.KubernetesConfig.EnablePodSecurityPolicy)
-		},
-		"OpenShiftGetMasterSh": func() (string, error) {
-			masterShAsset := getOpenshiftMasterShAsset(cs.Properties.OrchestratorProfile.OrchestratorVersion)
-			tb := MustAsset(masterShAsset)
-			t, err := template.New("master").Parse(string(tb))
-			if err != nil {
-				return "", err
-			}
-			b := &bytes.Buffer{}
-			err = t.Execute(b, struct {
-				ConfigBundle           string
-				ExternalMasterHostname string
-				RouterLBHostname       string
-				Location               string
-				MasterIP               string
-			}{
-				ConfigBundle:           base64.StdEncoding.EncodeToString(cs.Properties.OrchestratorProfile.OpenShiftConfig.ConfigBundles["master"]),
-				ExternalMasterHostname: fmt.Sprintf("%s.%s.cloudapp.azure.com", cs.Properties.MasterProfile.DNSPrefix, cs.Properties.AzProfile.Location),
-				RouterLBHostname:       fmt.Sprintf("%s-router.%s.cloudapp.azure.com", cs.Properties.MasterProfile.DNSPrefix, cs.Properties.AzProfile.Location),
-				Location:               cs.Properties.AzProfile.Location,
-				MasterIP:               cs.Properties.MasterProfile.FirstConsecutiveStaticIP,
-			})
-			return b.String(), err
-		},
-		"OpenShiftGetNodeSh": func(profile *api.AgentPoolProfile) (string, error) {
-			nodeShAsset := getOpenshiftNodeShAsset(cs.Properties.OrchestratorProfile.OrchestratorVersion)
-			tb := MustAsset(nodeShAsset)
-			t, err := template.New("node").Parse(string(tb))
-			if err != nil {
-				return "", err
-			}
-			b := &bytes.Buffer{}
-			err = t.Execute(b, struct {
-				ConfigBundle string
-				Role         api.AgentPoolProfileRole
-			}{
-				ConfigBundle: base64.StdEncoding.EncodeToString(cs.Properties.OrchestratorProfile.OpenShiftConfig.ConfigBundles["bootstrap"]),
-				Role:         profile.Role,
-			})
-			return b.String(), err
-		},
-		// inspired by http://stackoverflow.com/questions/18276173/calling-a-template-with-several-pipeline-parameters/18276968#18276968
-		"dict": func(values ...interface{}) (map[string]interface{}, error) {
-			if len(values)%2 != 0 {
-				return nil, errors.New("invalid dict call")
-			}
-			dict := make(map[string]interface{}, len(values)/2)
-			for i := 0; i < len(values); i += 2 {
-				key, ok := values[i].(string)
-				if !ok {
-					return nil, errors.New("dict keys must be strings")
-				}
-				dict[key] = values[i+1]
-			}
-			return dict, nil
-		},
-		"loop": func(min, max int) []int {
-			var s []int
-			for i := min; i <= max; i++ {
-				s = append(s, i)
-			}
-			return s
-		},
-		"subtract": func(a, b int) int {
-			return a - b
-		},
-		"IsCustomVNET": func() bool {
-			return isCustomVNET(cs.Properties.AgentPoolProfiles)
-		},
-	}
 }
 
 func makeMasterExtensionScriptCommands(cs *api.ContainerService) string {
