+++ pkg/acsengine/pki_test.go
@@ -103,7 +103,7 @@ func TestSubjectAltNameInCert(t *testing.T) {
 
 	// Test SAN is not empty
 	SubjectAltNames := []string{"santest.mydomain.com", "santest2.testdomain.net"}
-	masterExtraFQDNs := append(FormatAzureProdFQDNs(dnsPrefix), SubjectAltNames...)
+	masterExtraFQDNs := append(formatAzureProdFQDNs(dnsPrefix), SubjectAltNames...)
 	expectedDNSNames := []string{"santest.australiaeast.cloudapp.azure.com", "santest.australiasoutheast.cloudapp.azure.com", "santest.brazilsouth.cloudapp.azure.com", "santest.canadacentral.cloudapp.azure.com", "santest.canadaeast.cloudapp.azure.com", "santest.centralindia.cloudapp.azure.com", "santest.centralus.cloudapp.azure.com", "santest.centraluseuap.cloudapp.azure.com", "santest.chinaeast.cloudapp.chinacloudapi.cn", "santest.chinanorth.cloudapp.chinacloudapi.cn", "santest.eastasia.cloudapp.azure.com", "santest.eastus.cloudapp.azure.com", "santest.eastus2.cloudapp.azure.com", "santest.eastus2euap.cloudapp.azure.com", "santest.japaneast.cloudapp.azure.com", "santest.japanwest.cloudapp.azure.com", "santest.koreacentral.cloudapp.azure.com", "santest.koreasouth.cloudapp.azure.com", "santest.northcentralus.cloudapp.azure.com", "santest.northeurope.cloudapp.azure.com", "santest.southcentralus.cloudapp.azure.com", "santest.southeastasia.cloudapp.azure.com", "santest.southindia.cloudapp.azure.com", "santest.uksouth.cloudapp.azure.com", "santest.ukwest.cloudapp.azure.com", "santest.westcentralus.cloudapp.azure.com", "santest.westeurope.cloudapp.azure.com", "santest.westindia.cloudapp.azure.com", "santest.westus.cloudapp.azure.com", "santest.westus2.cloudapp.azure.com", "santest.chinaeast.cloudapp.chinacloudapi.cn", "santest.chinanorth.cloudapp.chinacloudapi.cn", "santest.germanycentral.cloudapp.microsoftazure.de", "santest.germanynortheast.cloudapp.microsoftazure.de", "santest.usgovvirginia.cloudapp.usgovcloudapi.net", "santest.usgoviowa.cloudapp.usgovcloudapi.net", "santest.usgovarizona.cloudapp.usgovcloudapi.net", "santest.usgovtexas.cloudapp.usgovcloudapi.net", "santest.francecentral.cloudapp.azure.com", "santest.mydomain.com", "santest2.testdomain.net", "kubernetes", "kubernetes.default", "kubernetes.default.svc", "kubernetes.default.svc.cluster.local", "kubernetes.kube-system", "kubernetes.kube-system.svc", "kubernetes.kube-system.svc.cluster.local"}
 
 	apiServerPair, _, _, _, _, _, err := CreatePki(masterExtraFQDNs, extraIPs, DefaultKubernetesClusterDomain, caPair, 1)
@@ -139,7 +139,7 @@ func TestSubjectAltNameInCert(t *testing.T) {
 
 	// Test SAN is empty
 	SubjectAltNames = []string{}
-	masterExtraFQDNs = append(FormatAzureProdFQDNs(dnsPrefix), SubjectAltNames...)
+	masterExtraFQDNs = append(formatAzureProdFQDNs(dnsPrefix), SubjectAltNames...)
 	expectedDNSNames = []string{"santest.australiaeast.cloudapp.azure.com", "santest.australiasoutheast.cloudapp.azure.com", "santest.brazilsouth.cloudapp.azure.com", "santest.canadacentral.cloudapp.azure.com", "santest.canadaeast.cloudapp.azure.com", "santest.centralindia.cloudapp.azure.com", "santest.centralus.cloudapp.azure.com", "santest.centraluseuap.cloudapp.azure.com", "santest.chinaeast.cloudapp.chinacloudapi.cn", "santest.chinanorth.cloudapp.chinacloudapi.cn", "santest.eastasia.cloudapp.azure.com", "santest.eastus.cloudapp.azure.com", "santest.eastus2.cloudapp.azure.com", "santest.eastus2euap.cloudapp.azure.com", "santest.japaneast.cloudapp.azure.com", "santest.japanwest.cloudapp.azure.com", "santest.koreacentral.cloudapp.azure.com", "santest.koreasouth.cloudapp.azure.com", "santest.northcentralus.cloudapp.azure.com", "santest.northeurope.cloudapp.azure.com", "santest.southcentralus.cloudapp.azure.com", "santest.southeastasia.cloudapp.azure.com", "santest.southindia.cloudapp.azure.com", "santest.uksouth.cloudapp.azure.com", "santest.ukwest.cloudapp.azure.com", "santest.westcentralus.cloudapp.azure.com", "santest.westeurope.cloudapp.azure.com", "santest.westindia.cloudapp.azure.com", "santest.westus.cloudapp.azure.com", "santest.westus2.cloudapp.azure.com", "santest.chinaeast.cloudapp.chinacloudapi.cn", "santest.chinanorth.cloudapp.chinacloudapi.cn", "santest.germanycentral.cloudapp.microsoftazure.de", "santest.germanynortheast.cloudapp.microsoftazure.de", "santest.usgovvirginia.cloudapp.usgovcloudapi.net", "santest.usgoviowa.cloudapp.usgovcloudapi.net", "santest.usgovarizona.cloudapp.usgovcloudapi.net", "santest.usgovtexas.cloudapp.usgovcloudapi.net", "santest.francecentral.cloudapp.azure.com", "kubernetes", "kubernetes.default", "kubernetes.default.svc", "kubernetes.default.svc.cluster.local", "kubernetes.kube-system", "kubernetes.kube-system.svc", "kubernetes.kube-system.svc.cluster.local"}
 
 	apiServerPair, _, _, _, _, _, err = CreatePki(masterExtraFQDNs, extraIPs, DefaultKubernetesClusterDomain, caPair, 1)
